Refactoring the folder structure in the code base
