+++ src/components/GalleryComponent.jsx
@@ -1,4 +1,3 @@
-// import { Card } from "@material-tailwind/react";
 import { useState } from "react";
 import PhotoAlbum from "react-photo-album";
 import Lightbox from "yet-another-react-lightbox";
