+++ src/components/about/about.js
@@ -1,0 +1,71 @@
+export const leftServices = [
+  `Inbound & Outbound Tours Travel`,
+  `Air Ticketing`,
+  `Air transfer`,
+  `Visa & Travel Insurance Processing`,
+  `Nature Expedition & People Cultural Programme`,
+  `Corporate Team Building`
+];
+
+export const rightServices = [
+  `Budget, Mid-range Safaris`,
+  `Hotel Booking`,
+  `Wildlife Safaris`,
+  `Beach Holidays`,
+  `Day Excursion Trips`,
+  `Honeymoon Packages`
+];
+
+export const values = [
+  {
+    title: 'Honesty',
+    body: `We are always ready to admit our actions
+    regardless of the situations we are in.`,
+  },
+  {
+    title: 'Transparency',
+    body: ` We have nothing to hide and are ready
+    to publicize our good performance as
+    wide as possible.`
+  },
+  {
+    title: 'Integrity',
+    body: `We beleive in being responsible.
+    We accomplish what we say we will do to our clients.`
+  },
+  {
+    title: 'Excellence',
+    body: `We try our best to make ourself so valuable in our work
+    so as to become indispensable.`
+  }
+]
+
+export const mission = [
+  {
+    title: 'Our Mission',
+    body: `We strive to be the premier travel service company committed to making
+    travel most pleasurable, safe, fun, exciting and hassle-free
+    for our customers through delivery of quality service that exceeds their expectations.`
+  },
+  {
+    title: 'Our Vision',
+    body: `Focusing on to be the leading tours and transport solution provider
+    and establish a professional managed travel and Tour Company of global standard.`
+  }
+]
+
+export const story = {
+  title: 'Our Stroy',
+  body: `Bush and Beach Travel & Safaris Ltd is a fully integrated tour and travels
+  company with a proven track record that provides customised luxury
+  journeys to inspiring destinations.We are a trusted company, passionate travel
+  specialists who strive to ensure that we are and remain the perfect solution to transport
+  and tour requirements in East Africa.Our trips are completely based around
+  our clients’ interests and schedule but also specialist in one-of-a-kind
+  beach experiences and expeditions.At bush and beach we provide the most magical opportunities to
+  visit and see the discovered and undiscovered places in Kenya,
+  Africa and across the globe. From the highlands of Mt Kenya
+  to shores of the Indian Ocean, from Egyptian Pyramids in
+  Cape Town to the Eiffel Tower in France, Bush and Beach
+  Travel & Safaris Ltd believes in the best experience of a Safari.`
+}
