+++ src/components/bangkok/data.js
@@ -36,3 +36,61 @@ export const thailandImages = [
     alt: "Thailand"
   }
 ]
+
+export const bangkokItenary = [
+  {
+    day: `Day 1`,
+    title: `Arrival in Bangkok, Thailand`,
+    activities: [
+      `Begin your Thai adventure in the vibrant city of Bangkok. Upon arrival, our team will welcome you and transfer you to your comfortable accommodation. Feel the buzz of the city and prepare for an unforgettable journey.`
+    ]
+  },
+  {
+    day: `Day 2`,
+    title: `Bangkok City Exploration`,
+    activities: [
+      `Delve into Bangkok's cultural and historical treasures. Visit the Grand Palace, Wat Phra Kaew (Temple of the Emerald Buddha), and take a boat ride along the Chao Phraya River. Immerse yourself in the city's captivating blend of tradition and modernity.`,
+    ]
+  },
+  {
+    day: `Day 3`,
+    title: `Discover Bangkok's Charms`,
+    activities: [
+      `Embark on a day of exploration. Visit the bustling Damnoen Saduak Floating Market, where you can experience the vibrant local culture and sample delicious street food.`,
+    ]
+  },
+  {
+    day: `Day 4`,
+    title: `Travel to Phuket`,
+    activities: [
+      `Fly to the tropical paradise of Phuket. Settle into your beachfront accommodation and unwind under the palm trees. Enjoy the pristine beaches, turquoise waters, and laid-back island vibes.`
+    ]
+  },
+  {
+    day: `Day 5`,
+    title: `Phuket Island Adventures`,
+    activities: [
+      `Discover the natural wonders of Phuket. Explore Phi Phi Islands, snorkel through crystal-clear waters, and experience the vibrant marine life that makes these islands a paradise for underwater enthusiasts.`
+    ]
+  },
+  {
+    day: `Day 6`,
+    title: `Beach Bliss in Phuket`,
+    activities: [
+      `Spend your day soaking up the sun and enjoying the island's leisure activities. Engage in water sports, explore local markets, or simply relax on the sandy shores.`
+    ]
+  },
+  {
+    day: `Day 7`,
+    title: `Reflection and Departure`,
+    activities: [
+      `Savor your last moments in Thailand. Depending on your flight time, you may have the opportunity for a final exploration or relaxation. Depart with cherished memories of your Bangkok and Phuket escapade.`,
+    ]
+  },
+]
+export const bangkokPromoAd = {
+  promoTitle: `Unveil the excitement of Bangkok's city life and the tranquility of Phuket's beaches.`,
+  promoBody: `Unveil the beauty and diversity of Paris, Barcelona, and Rome. Limited slots are available for this exclusive offer.`,
+  price: `$1875pp`,
+  callToAction: `Contact us now to secure your spot on a journey that promises a perfect balance of adventure and relaxation!`
+};
