+++ src/components/china/data.js
@@ -21,3 +21,55 @@ export const chinaImages = [
     alt: "China"
   }
 ]
+
+export const chinaItenary = [
+  {
+    day: `Day 1`,
+    title: `Arrival in Guangzhou, China`,
+    activities: [
+      `Welcome to the bustling city of Guangzhou! Upon arrival, our team will ensure a smooth transfer to your comfortable accommodation. Rest up for the days of business and exploration ahead.`
+    ]
+  },
+  {
+    day: `Day 2`,
+    title: `Canton Fair Trade Exhibition Begins`,
+    activities: [
+      `Participate in the prestigious Canton Fair, where business opportunities and connections await. Engage with exhibitors, discover new products, and build valuable partnerships. Immerse yourself in the world of trade.`
+    ]
+  },
+  {
+    day: `Day 3-7`,
+    title: `Canton Fair Business Days`,
+    activities: [
+      `Dive into the business days of the Canton Fair. Network, negotiate deals, and explore the vast array of industries represented at this renowned trade exhibition.`
+    ]
+  },
+  {
+    day: `Day 8`,
+    title: `Explore Guangzhou`,
+    activities: [
+      `As the Canton Fair concludes, take the day to explore Guangzhou's cultural attractions. Visit the Chen Clan Ancestral Hall, explore the Shamian Island, and savor authentic Cantonese cuisine.`
+    ]
+  },
+  {
+    day: `Day 9`,
+    title: `Day Trip to Foshan`,
+    activities: [
+      `Embark on a day trip to Foshan, a city known for its rich history in arts and crafts. Visit the Foshan Ancestral Temple and experience the city's cultural heritage.`
+    ]
+  },
+  {
+    day: `Day 10`,
+    title: `Departure`,
+    activities: [
+      `Reflect on your Canton Fair experience and bid farewell to China. Depending on your flight time, you may have the opportunity for some last-minute exploration or shopping. Depart with new business connections and memories.`
+    ]
+  },
+]
+
+export const chinaPromoAd = {
+  promoTitle: `Book Your Canton Fair Experience Today!`,
+  promoBody: `Join the bustling world of trade at the Canton Fair and delve into the cultural richness of Guangzhou and Foshan.`,
+  price: `$1890pp`,
+  callToAction: `Contact us now to secure your spot on a journey that combines business and exploration in China!`
+};
