+++ src/components/dubai/data.js
@@ -26,3 +26,49 @@ export const dubaiImages = [
     alt: "Shopping mall"
   },
 ];
+
+export const dubai4NightsItenary = [
+  {
+    day: `Day 1`,
+    title: `Arrival in Dubai`,
+    activities: [
+      `Arrive in the dazzling city of Dubai, where modern marvels meet rich tradition. Our representative will greet you and transfer you to your luxurious accommodation. Take the evening to unwind and marvel at the city's iconic skyline.`
+    ]
+  },
+  {
+    day: `Day 2`,
+    title: `City Exploration and Desert Safari`,
+    activities: [
+      `Embark on a guided city tour, visiting landmarks like the Burj Khalifa, Dubai Mall, and the Palm Jumeirah. In the afternoon, get ready for an exhilarating desert safari. Dune bashing, camel rides, and a stunning desert sunset await.`
+    ]
+  },
+  {
+    day: `Day 3`,
+    title: `Luxury and Shopping Extravaganza`,
+    activities: [
+      `Indulge in a day of shopping and relaxation. Visit the renowned Dubai Mall for high-end shopping and experience the thrill of the indoor Dubai Aquarium. Unwind in your hotel's spa or take a leisurely stroll along Jumeirah Beach.`
+    ]
+  },
+  {
+    day: `Day 4`,
+    title: `Abu Dhabi Excursion`,
+    activities: [
+      `Take a day trip to Abu Dhabi, the UAE's capital. Explore the majestic Sheikh Zayed Grand Mosque, the stunning Emirates Palace, and the futuristic Louvre Abu Dhabi. Return to Dubai in the evening, where the city's lights will dazzle you.`
+    ]
+  },
+  {
+    day: `Day 5`,
+    title: `Farewell to Dubai`,
+    activities: [
+      `Before departure, visit the historic Dubai Creek and engage in a traditional abra boat ride. Capture a few last-minute memories before bidding farewell to this extraordinary city.`
+    ]
+  },
+]
+
+export const dubai4NightsPromoAd = {
+  promoTitle: `Book Your Dubai Getaway Today!`,
+  promoBody: `Experience the magic of Dubai's skyscrapers, deserts, and luxury. Limited slots are available for this exclusive offer.`,
+  price: `$650pp`,
+  callToAction: `Contact us to embark on a journey that 
+  will redefine your connection with nature and wildlife.`
+};
