+++ src/components/egpyt-israel/data.js
@@ -31,3 +31,70 @@ export const egyptIsraelImages = [
     alt: "Egyp and Israel"
   }
 ]
+
+export const egyptIsraelItenary = [
+  {
+    day: `Day 1`,
+    title: `Arrival in Cairo, Egypt`,
+    activities: [
+      `Begin your journey of faith in Cairo, Egypt's capital. Our warm welcome awaits as we transfer you to your comfortable accommodation. Take the evening to rest and reflect on the spiritual adventure ahead.`
+    ]
+  },
+  {
+    day: `Day 2`,
+    title: `Explore Ancient Egypt`,
+    activities: [
+      `Embark on a soul-enriching day of exploration. Visit the Giza Pyramids, Sphinx, and the Egyptian Museum, where artifacts from ancient civilizations await. Reflect on the connections between history and faith.`
+    ]
+  },
+  {
+    day: `Day 3`,
+    title: `Journey to Mount Sinai`,
+    activities: [
+      `Travel to the sacred Mount Sinai, where Moses received the Ten Commandments. Ascend the mountain to witness the sunrise, a moment of spiritual significance for pilgrims.`
+    ]
+  },
+  {
+    day: `Day 4`,
+    title: `Arrive in Jerusalem, Israel`,
+    activities: [
+      `Cross into Israel and arrive in Jerusalem, the heart of spirituality. Settle into your accommodation and prepare to explore the revered sites that hold deep meaning for multiple faiths.`
+    ]
+  },
+  {
+    day: `Day 5`,
+    title: `Jerusalem's Holy Sites`,
+    activities: [
+      `Embark on a profound day of visiting sacred sites. Explore the Western Wall, Church of the Holy Sepulchre, and the Dome of the Rock, reflecting on the spiritual heritage shared by many.`
+    ]
+  },
+  {
+    day: `Day 6`,
+    title: `Bethlehem and Jericho`,
+    activities: [
+      `Visit Bethlehem, the birthplace of Jesus, and explore the Church of the Nativity. Continue to Jericho, one of the world's oldest cities, and experience its historical and spiritual significance.`
+    ]
+  },
+  {
+    day: `Day 7`,
+    title: `Journey to Galilee`,
+    activities: [
+      `Travel to the serene region of Galilee, where Jesus performed many miracles. Visit sites like the Mount of Beatitudes, Tabgha, and Capernaum, connecting with the teachings of Christ.`
+    ]
+  },
+  {
+    day: `Day 8`,
+    title: `Reflection and Departure`,
+    activities: [
+      `Take a day of reflection and prayer. Depending on your flight time, you may have the opportunity to revisit a site or simply absorb the spiritual essence of your pilgrimage. Depart with a heart full of faith and memories.
+      `,
+    ]
+  },
+]
+
+export const egyptIsraelPromoAd = {
+  promoTitle: `Book Your Pilgrimage Journey Today!`,
+  promoBody: `Experience the sacredness of Egypt and Israel's sites, tracing the footsteps of prophets and connecting with centuries of faith.`,
+  price: `$2450pp`,
+  callToAction: `Contact us now to secure your spot on a journey that will deepen your spiritual connection and understanding.`
+};
