+++ src/components/international/international.js
@@ -1,0 +1,139 @@
+import dubai from '../../assets/images/cities/dubai.jpg';
+import joburg from '../../assets/images/cities/capetown.jpg'
+import thailand from '../../assets/images/cities/thailand.jpg';
+import turkey from '../../assets/images/cities/turkey.jpg';
+import rome from '../../assets/images/cities/rome.jpg';
+import mauritius from '../../assets/images/cities/mauritius.jpg';
+import china from '../../assets/images/cities/china.jpg';
+import singapore from '../../assets/images/cities/singapore.jpg';
+import israel from '../../assets/images/cities/israel.jpg';
+
+// import dubai from '../assets/images/cities/dubai.jpg';
+// import joburg from '../assets/images/cities/capetown.jpg'
+// import thailand from '../assets/images/cities/thailand.jpg';
+// import turkey from '../assets/images/cities/turkey.jpg';
+// import rome from '../assets/images/cities/rome.jpg';
+// import mauritius from '../assets/images/cities/mauritius.jpg';
+// import china from '../assets/images/cities/china.jpg';
+// import singapore from '../assets/images/cities/singapore.jpg';
+// import israel from '../assets/images/cities/israel.jpg';
+
+export const cities = [dubai, joburg, singapore, mauritius, china, israel, rome, thailand, turkey]
+
+export const intlTourDestinations = [
+  {
+    title: "Discover Dubai",
+    image: dubai,
+    days: "5",
+    nights: "4",
+    description: `Our meticulously crafted itinerary ensures you experience the city's highlights. All starting from $650 per person, this offer provides an unbeatable value for an unforgettable experience`,
+    url: "dubai",
+  },
+  {
+    title: "Johannesburg & Sun City Delight",
+    description: `Embark on an unforgettable 7-day journey through the dynamic city of Johannesburg and the luxurious oasis of Sun City. Starting from just $1495 per person.`,
+    image: joburg,
+    days: "7",
+    nights: "6",
+    url: "joburg",
+  },
+  {
+    title: "Malaysia & Singapore Exploration",
+    image: singapore,
+    days: "7",
+    nights: "6",
+    description: `Embark on a captivating 7-day journey through the cultural richness of Kuala Lumpur and the modern allure of Singapore. Starting from just $2310 per person`,
+    url: "malaysia",
+  },
+  {
+    title: "Mauritius Escapade",
+    image: mauritius,
+    days: "4",
+    nights: "3",
+    description: `Experience the paradisiacal beauty of Mauritius in just 5 days! Starting from $1700 per person, this offer ensures you experience the best of the island's beaches, culture, and natural wonders.`,
+    url: "mauritius",
+  },
+  {
+    title: "China Canton Fair Trade Exhibition Adventure",
+    image: china,
+    days: "10",
+    nights: "9",
+    description: `Starting from $1890 per person, you ge to engage in the dynamic world of trade at the Canton Fair Exhibition as well as an opportunity to explore Chinese culture.`,
+    url: `china`,
+    
+  },
+  {
+    title: "Sacred Pilgrimage Tour: Egypt & Israel Journey of Faith",
+    image: israel,
+    days: "9",
+    nights: "8",
+    description: `Embark on a spiritual pilgrimage through Egypt and Israel, exploring sacred sites and connecting with faith. Starting from just $2450 per person.`,
+    url: "egypt-israel",
+  },
+  {
+    title: "Paris, Barcelona & Italy Adventure",
+    image: rome,
+    days: "7",
+    nights: "6",
+    description: `Experience the allure of three European gems in just 7 days! Starting from $1790 per person, this offer provides you with a taste of Parisian fashion, Barcelona's vibrancy, and the historical wonders of Rome.`,
+    url: "italy-barca-paris",
+  },
+  {
+    title: "Bangkok & Phuket Getaway: 7 Days & 6 Nights of Thai Adventure!",
+    image: thailand,
+    days: "7",
+    nights: "6",
+    description: `Get to Experience the best of both worlds in 
+    Thailand's bustling capital and tropical paradise for just $1875 per person. 
+    Experience the beauty, culture, 
+    and relaxation Thailand has to offer.`,
+    url: "bangkok",
+  },
+  {
+    title: "Istanbul Discovery: 5 Days & 4 Nights Journey through Turkey's Cultural Heart",
+    image: turkey,
+    days: "5",
+    nights: "4",
+    description: `Embark on a journey through the heart of Turkey's cultural gem, 
+    Istanbul. Starting from just $1660 per person, you experience the city's rich history, 
+    vibrant markets, and iconic landmarks.`,
+    url: "turkey"
+  },
+]
+
+export const title = `Adventure like no other.`;
+
+export const bodyLeft = `
+      Welcome to Kenya which is known for being the world’s best safari
+      destination with more than 50 epic national parks and reserves
+      that are home to diverse wildlife including the big five
+      (Lions, Leopards, Elephants, Rhinoceros and Buffalos).
+      You will get to experience the notorious wildebeest migration
+      with a picturesque view of over 2 million animals migrating from
+      Serengeti national park into Masai Mara national park which normally
+      happens in July to October every year. This experience has been
+      ranked as one of the seven wonders of the world.
+    `;
+export const bodyRight = `
+      Kenya is an ideal place for visitors to enjoy breath-taking tourism safaris.
+      There are several activities including the hot air balloon safaris conducted
+      majority in Masai Mara and Amboseli national reserve making it an unforgettable
+      experience to view wildlife from the air. The organized safaris
+      will help you visit several destinations in Kenya in the shortest
+      time possible. Our Kenyan bush safaris is categorized as below
+      depending on how long you would wish to stay enabling you to have
+      a captivating and a memorable experience;
+    `;
+
+export const title2 = `A Symphony Of Untamed Nature's Beauty.`;
+
+export const bodyTextTwo = `
+      Let us take you from the iconic savannahs of the Maasai Mara, where golden grasses sway
+      in harmony with the rhythms of the wild, to the enchanting embrace of
+      Amboseli&apos;s elephants against the backdrop of towering Kilimanjaro, and also to witness
+      the great wildebeest migration which paints the plains with a
+      mesmerizing spectacle of life&apos;s eternal cycle.
+      Come with us to see the Lions, leopards, and cheetahs prowl
+      with regal grace both in the city as well as in the the grasslands.
+      Come with us to Lake Nakuru to witness the theater of flamingo-pink ballet.
+    `;
