+++ src/components/joburg/data.js
@@ -21,3 +21,63 @@ export const joburgImages = [
     alt: "Johanesburg"
   }
 ]
+
+// Joburg 
+export const joburgItenary = [
+  {
+    day: `Day 1`,
+    title: `Arrival in Johannesburg`,
+    activities: [
+      `Arrive in the vibrant city of Johannesburg, your gateway to South African wonders. Our warm welcome awaits as we transfer you to your comfortable accommodation. Spend the evening relaxing and anticipating the exciting days ahead.`
+    ]
+  },
+  {
+    day: `Day 2`,
+    title: `Johannesburg Exploration`,
+    activities: [
+      `Embark on a city tour, delving into Johannesburg's rich history and culture. Visit the Apartheid Museum, Constitution Hill, and explore the vibrant neighborhoods that reflect the city's diversity.`
+    ]
+  },
+  {
+    day: `Day 3`,
+    title: `Sun City Arrival`,
+    activities: [
+      `Travel to the luxurious Sun City resort, an oasis of entertainment and relaxation. Check in to your resort and immerse yourself in the lush surroundings, pristine pools, and exhilarating activities.`
+    ]
+  },
+  {
+    day: `Day 4`,
+    title: `Sun City Adventures`,
+    activities: [
+      `Enjoy a day of endless possibilities in Sun City. From thrilling water activities at the Valley of Waves to exploring the exquisite gardens and entertainment complexes, there's something for everyone.`
+    ]
+  },
+  {
+    day: `Day 5`,
+    title: `Pilanesberg National Park Safari`,
+    activities: [
+      `Embark on a safari adventure in Pilanesberg National Park, adjacent to Sun City. Encounter Africa's iconic wildlife, including the Big Five, on a thrilling game drive.`
+    ]
+  },
+  {
+    day: `Day 6`,
+    title: `Return to Johannesburg`,
+    activities: [
+      `Bid farewell to Sun City and return to Johannesburg. Spend the afternoon at leisure, perhaps indulging in some last-minute shopping or exploring more of the city's attractions.`
+    ]
+  },
+  {
+    day: `Day 7`,
+    title: `Departure`,
+    activities: [
+      `Before departure, take a moment to reflect on your South African adventure. Depending on your flight time, you may have the opportunity for a final exploration or relaxation. Depart with cherished memories and the promise of returning.`
+    ]
+  },
+]
+
+export const joburgPromoAd = {
+  promoTitle: `Book Your South African Adventure Today!`,
+  promoBody: `Discover the vibrant city life of Johannesburg and the leisurely paradise of Sun City. Limited slots are available for this exclusive offer. `,
+  price: `$1495pp`,
+  callToAction: `Contact us now to secure your spot on a journey that will immerse you in the splendor and diversity of South Africa's attractions.`
+};
