+++ src/components/kenya-safari/amboseli-tsavo-west/data.js
@@ -1,0 +1,48 @@
+export const amboseliTsavoItenary = [
+  {
+    day: `Day 1`,
+    title: `Journey to Amboseli's Majesty`,
+    activities: [
+      `Commence your adventure as we whisk you away to Amboseli National Park.`,
+      `et against the backdrop of Mount Kilimanjaro, experience breathtaking views and remarkable elephant herds.`,
+      `Engage in an afternoon game drive, capturing the essence of Amboseli's unique charm.`,
+    ]
+  },
+  {
+    day: `Day 2`,
+    title: `Amboseli's Enchanting Wildlife`,
+    activities: [
+      `Spend a full day exploring Amboseli's diverse landscapes and its abundant wildlife.`,
+      `Lions, cheetahs, and a plethora of bird species await your discovery`,
+      `Capture the iconic elephants with Kilimanjaro's snow-capped peak as a backdrop.`,
+    ]
+  },
+  {
+    day: `Day 3`,
+    title: `Tsavo West's Hidden Gems`,
+    activities: [
+      `Embark on a scenic journey to Tsavo West National Park, a realm of hidden treasures.`,
+      `Venture into the Mzima Springs to observe hippos and crocodiles underwater.`,
+      `Explore the Shetani Lava Flows and marvel at the ancient landscapes of this unique wilderness`
+    ]
+  },
+  {
+    day: `Day 4`,
+    title: `Reflecting on Nature's Grandeur`,
+    activities: [
+      `Conclude your journey with a morning game drive in Tsavo West`,
+      `Relish the final moments with the wildlife and landscapes that have left an indelible mark on your heart`,
+      `Return to your starting point with memories of Kenya's untamed beauty.`
+    ]
+  },
+]
+
+export const amboseliTsavoPromoAd = {
+  promoTitle: `Capture the Essence of Kenya's Wilderness!`,
+  promoBody: `Embark on a captivating 4-day journey through Amboseli 
+  and Tsavo West's natural wonders. `,
+  price: `Ksh.15,000`,
+  callToAction: `Contact us now to secure your spot 
+  on an unforgettable safari experience that unveils the 
+  beauty of Kenya's landscapes and wildlife!`
+};
