+++ src/components/kenya-safari/lake-nakuru/LakeNakuruMara.jsx
@@ -1,7 +1,7 @@
-import SafariDetail from "./Safari";
-import Footer from "../Footer";
-import { maraImages } from "../../data/safari";
-import { lakeNakuruIternary, lakeNakPromoAd } from "../../data/itenaries";
+import SafariDetail from "../../modals/Safari";
+import Footer from "../../Footer";
+import { maraImages } from "../../../data/safari";
+import { lakeNakPromoAd, lakeNakuruIternary } from "./data";
 const LakeNakuruMaraTrip = () => (
   <>
     <SafariDetail
