+++ src/components/kenya-safari/lake-nakuru/data.js
@@ -1,0 +1,52 @@
+export const lakeNakuruIternary = [
+  {
+    day: `Day 1`,
+    title: 'Lake Nakuru Welcomes You',
+    activities: [
+      `Depart from Nairobi and journey to Lake Nakuru`,
+      `Marvel at the flamingo spectacle and diverse birdlife.`,
+      `Thrilling game drive to spot rhinos, lions, and more.`,
+      `Overnight stay in a luxurious lodge by the lake.`
+    ]
+  },
+  {
+    day: `Day 2`,
+    title: `Into the Heart of Masai Mara`,
+    activities: [
+      `Travel to the enchanting Masai Mara.`,
+      `Afternoon game drive in search of the Big Five`,
+      `Immerse yourself in the untamed beauty of the Mara.`,
+      `Unwind at our carefully selected lodge.`,
+    ]
+  },
+  {
+    day: `Day 3`,
+    title: 'The Great Migration Experience',
+    activities: [
+      `Embark on a full-day game drive to witness the Great Migration.`,
+      `Capture unforgettable moments as herds traverse the plains.`,
+      `Encounter predators and their thrilling pursuits.`,
+      `Evening relaxation and overnight stay in Masai Mara.`,
+    ]
+  },
+  {
+    day: `Day 4`,
+    title: 'Farewell to the Wilderness',
+    activities: [
+      `Capture final glimpses of Mara's wildlife on a morning game drive.`,
+      `Return to Nairobi, cherishing memories of a lifetime.`,
+      `Your adventure concludes with drop-off at your preferred location.`,
+    ],
+  },
+];
+
+export const lakeNakPromoAd = {
+  promoTitle: `Book Now for an Unforgettable Safari!`,
+  promoBody: `Don't miss this chance to explore two of Kenya's most iconic destinations. 
+  Immerse yourself in the beauty of Lake Nakuru and witness 
+  the grandeur of the Masai Mara. Limited slots available.`,
+  price: `Ksh.15,000`,
+  callToAction: `Contact us to secure your spot on 
+  this extraordinary 4-day journey into the wild!
+  `
+};
