+++ src/components/kenya-safari/masaai-mara/MasaaiMara.jsx
@@ -1,7 +1,7 @@
-import SafariDetail from "./Safari";
-import Footer from "../Footer";
-import { maraImages } from "../../data/safari";
-import { masaaiMaraIternary, maraPromoAd } from "../../data/itenaries";
+import SafariDetail from "../../modals/Safari";
+import Footer from "../../Footer";
+import { maraImages } from "../../../data/safari";
+import { masaaiMaraIternary, masaaiMaraPromoAd } from "./data";
 const MasaaiMaraTrip = () => (
   <>
     <SafariDetail
@@ -9,7 +9,7 @@ const MasaaiMaraTrip = () => (
       duration="3 Days & 2 Nights"
       safariTitle="Masaai Mara"
       itenary={masaaiMaraIternary}
-      formData={maraPromoAd}
+      formData={masaaiMaraPromoAd}
      />
     <Footer />
   </>
