+++ src/components/kenya-safari/masaai-mara/data.js
@@ -1,0 +1,42 @@
+export const masaaiMaraIternary = [
+  {
+    day: `Day 1`,
+    title: 'Journey to Masai Mara',
+    activities: [
+      `Embark on a thrilling adventure as we transport you to the iconic Masai Mara.`,
+      `Feel the anticipation build as you approach this legendary wilderness.`,
+      `Settle into your luxurious lodge, surrounded by the sights and sounds of the African savanna.`
+    ],
+  },
+  {
+    day: `Day 2`,
+    title: `Unveiling the Mara's Wonders`,
+    activities: [
+      `Rise with the sun for a day of unparalleled discovery.
+       Engage in heart-pounding game drives as the Mara's wildlife awakens
+       Witness lions on the prowl, elephants at play, and graceful giraffes on the horizon.
+       Experience the untamed beauty that has captivated explorers for generations.`,
+    ],
+  },
+
+  {
+    day: `Day 3`,
+    title: 'Farewell to the Mara',
+    activities: [
+      `Capture final glimpses of the Mara's wildlife on a morning game drive, etching these memories in your heart,
+      As the sun sets on your journey, return to Nairobi with a profound connection to the wild,
+      Your adventure concludes with drop-off at your preferred location.`,
+    ],
+  },
+];
+
+export const masaaiMaraPromoAd = {
+  promoTitle: `Capture the Essence of Africa's Wilderness!`,
+  promoBody: `Experience the awe-inspiring landscapes and incredible wildlife 
+  of Masai Mara in just 3 days. `,
+  price: `Ksh.15,000`,
+  callToAction: `
+  Contact us now to secure your spot on a journey 
+  that promises unforgettable 
+  moments in the heart of the wild!`,
+}
