+++ src/components/kenya-safari/nrb-amboseli-lewa-mara/NairobiAmboseliLewaMara.jsx
@@ -1,7 +1,7 @@
-import SafariDetail from "./Safari";
-import Footer from "../Footer";
-import { maraImages } from "../../data/safari";
-import { nrbAmboseliLewaItenary, nrbAmboseliLewaPromo } from "../../data/itenaries";
+import SafariDetail from "../../modals/Safari";
+import Footer from "../../Footer";
+import { maraImages } from "../../../data/safari";
+import { nrbAmboseliLewaItenary, nrbAmboseliLewaPromo } from "./data";
 
 const NairobiAmboseliLewaMaraTrip = () => (
   <>
