+++ src/components/kenya-safari/nrb-amboseli-lewa-mara/data.js
@@ -1,0 +1,58 @@
+export const nrbAmboseliLewaItenary = [
+  {
+    day: `Day 1`,
+    title: `Arrival in Nairobi`,
+    activities: [
+      `Begin your adventure in Nairobi, Kenya's dynamic capital.`,
+      `Immerse yourself in its urban culture, explore local markets, or visit wildlife sanctuaries.`,
+      `Settle into your city accommodation and anticipate the enchanting journey ahead.`,
+    ]
+  },
+  {
+    day: `Day 2 - 3`,
+    title: `Amboseli's Grace and Majesty`,
+    activities: [
+      `Embark to Amboseli National Park, where the magnificence of Mount Kilimanjaro unfolds.`,
+      `Engage in exhilarating game drives to witness elephants, lions, and the park's iconic landscapes.`,
+      `Capture unforgettable moments under Africa's tallest peak.`,
+    ]
+  },
+  {
+    day: `Day 4 - 5`,
+    title: `Lewa Conservancy's Wilderness`,
+    activities: [
+      `Journey to the Lewa Conservancy, a private haven dedicated to conservation.`,
+      `Embark on game drives, encountering rare species such as Grevy's zebras and black rhinos.`,
+      `Engage in conservation activities and experience the untamed beauty of this exclusive sanctuary.`,
+    ]
+  },
+  {
+    day: `Day 6 - 7`,
+    title: `Majestic Masai Mara Awaits`,
+    activities: [
+      `Travel to the renowned Masai Mara for two days of remarkable adventure.`,
+      `Witness the Great Migration, spot the Big Five, and capture the essence of the African wilderness on exhilarating game drives.`,
+      `Engage with Maasai communities and relish the Mara's captivating sunsets.`
+    ]
+  },
+  {
+    day: `Day 8`,
+    title: `Reflections on a Multifaceted Journey`,
+    activities: [
+      `Conclude your safari with a final game drive in Masai Mara.`,
+      ` Return to Nairobi, carrying cherished memories of diverse landscapes, captivating wildlife, and rich cultural encounters.`,
+    ]
+  },
+]
+
+export const nrbAmboseliLewaPromo = {
+  promoTitle: `Capture the Essence of Kenya's Diversity and Conservation Efforts!`,
+  promoBody: `
+  Embark on an extraordinary 8-day journey, 
+  traversing Kenya's diverse landscapes, 
+  rich cultures, and incredible wildlife encounters.`,
+  price: `Ksh.15,000`,
+  callToAction: `Contact us now to secure your spot on a once-in-a-lifetime exploration 
+  that encompasses Nairobi, Amboseli, 
+  Lewa Conservancy, and the captivating Masai Mara! `
+};
