+++ src/components/kenya-safari/nrb-amboseli-mara-mt-kenya/data.js
@@ -1,0 +1,68 @@
+export const nrbAmboseliMtKenyaItenary = [
+  {
+    day: `Day 1`,
+    title: `Welcome to Nairobi`,
+    activities: [
+      `Arrive in Nairobi, Kenya's bustling capital.`,
+      `Explore its vibrant markets, museums, and wildlife centers.`,
+      `Settle into your city accommodation and prepare for the extraordinary journey ahead.`,
+    ]
+  },
+  {
+    day: `Day 2 - 3`,
+    title: `Amboseli's Majestic Charm`,
+    activities: [
+      `Head to Amboseli National Park, where the mighty Mount Kilimanjaro serves as the backdrop.`,
+      `Engage in exhilarating game drives to witness elephants, lions, and a diversity of species.`,
+      `Relish the magic of the African wilderness.`,
+    ]
+  },
+  {
+    day: `Day 4`,
+    title: `Exploring the Mystical Mt. Kenya`,
+    activities: [
+      `Travel to the awe-inspiring Mount Kenya region.`,
+      `Immerse yourself in the lush landscapes, witness diverse flora and fauna, and take in the serenity of the highland environment.`,
+      `Overnight stay near the majestic mountain.`,
+    ]
+  },
+  {
+    day: `Day 5`,
+    title: `Lake Nakuru's Pink Spectacle`,
+    activities: [
+      `Journey to Lake Nakuru National Park, a paradise for bird enthusiasts`,
+      `Witness the breathtaking sight of flamingos painting the lake pink.`,
+      `Encounter rhinos, lions, and other wildlife amidst stunning vistas.`
+    ]
+  },
+  {
+    day: `Day 6-7`,
+    title: `Majestic Masai Mara Awaits`,
+    activities: [
+      `Travel to the iconic Masai Mara for two days of unparalleled adventure.`,
+      `Witness the Great Migration, spot the Big Five, and capture unforgettable moments on game drives.`,
+      `Engage with Maasai communities and savor the African sunset.`
+    ]
+  },
+  {
+    day: 'Day 8',
+    title: `Reflections on an Epic Journey`,
+    activities: [
+      `Conclude your safari with a final game drive in Masai Mara.`,
+      `Return to Nairobi, carrying memories of diverse landscapes, 
+      captivating wildlife, and rich cultural encounters`,
+    ]
+
+  },
+]
+
+export const nrbAmboseliMtKenyaPromoAd = {
+  promoTitle: `Capture the Essence of Kenya's Diversity and Splendor!`,
+  promoBody: `
+  Embark on the journey of a lifetime, spanning Kenya's iconic 
+  landscapes, vibrant cultures, and mesmerizing wildlife encounters.`,
+  price: `Ksh.15,000`,
+  callToAction: `Contact us now to secure your place on an unforgettable 
+  safari that explores Nairobi, Amboseli, Mt. Kenya, Lake Nakuru, 
+  and the captivating Masai Mara!`
+};
