+++ src/components/kenya-safari/nrb-amboseli-mara/NairobiAmboseliMara.jsx
@@ -1,7 +1,7 @@
-import SafariDetail from "./Safari";
-import Footer from "../Footer";
-import { maraImages } from "../../data/safari";
-import { nrbAmboseliMaraItenary, nrbAmoseliMaraPromo } from "../../data/itenaries";
+import SafariDetail from "../../modals/Safari";
+import Footer from "../../Footer";
+import { maraImages } from "../../../data/safari";
+import { nrbAmboseliMaraItenary, nrbAmoseliMaraPromo } from "./data";
 
 const NairobiAmboseliMaraTrip = () => (
   <>
