+++ src/components/kenya-safari/nrb-amboseli-mara/data.js
@@ -1,0 +1,49 @@
+export const nrbAmboseliMaraItenary = [
+  {
+    day: `Day 1`,
+    title: 'Arrival in Nairobi',
+    activities: [
+      `Your journey begins in Nairobi, Kenya's dynamic capital.`,
+      `Explore its vibrant culture, visit the David Sheldrick Wildlife Trust, or indulge in local markets`,
+      `Settle into your city accommodation and prepare for the diverse adventure ahead.`,
+    ]
+  },
+  {
+    day: `Day 2 - 3`,
+    title: `Amboseli's Majesty`,
+    activities: [
+      `Venture to Amboseli National Park, an iconic realm beneath the gaze of Mount Kilimanjaro.`,
+      `Immerse yourself in unrivaled landscapes and experience thrilling game drives.`,
+      `Encounter herds of elephants against the backdrop of Africa's tallest peak.`,
+    ]
+  },
+  {
+    day: `Day 4 - 5`,
+    title: 'The Enchanting Masai Mara',
+    activities: [
+      `Travel to the world-famous Masai Mara, where two days of unforgettable experiences await. `,
+      `Witness the grandeur of the Great Migration and spot the Big Five on exhilarating game drives`,
+      `Engage with Maasai communities and soak in the magic of the Mara's sunsets.`,
+    ]
+  },
+  {
+    day: `Day 6`,
+    title: 'Reflecting on an Extraordinary Journey',
+    activities: [
+      `Conclude your adventure with a morning game drive in Masai Mara, 
+      capturing last memories of the wildlife.`,
+      `Return to Nairobi, cherishing the encounters and landscapes that have left an indelible mark on your heart.`,
+    ],
+  },
+
+];
+
+export const nrbAmoseliMaraPromo = {
+  promoTitle: `Capture the Essence of Kenya's Diversity! `,
+  promoBody: `Embark on an all-encompassing 6-day expedition, 
+  spanning city life, breathtaking landscapes, 
+  and remarkable wildlife encounters.`,
+  price: `Ksh.15,000`,
+  callToAction: `Contact us now to secure your place on an unforgettable 
+  safari adventure from Nairobi to Amboseli and the enchanting Masai Mara!`
+};
