+++ src/components/kenya-safari/nrb-samburu-mara/NairobiSamburuMara.jsx
@@ -1,7 +1,8 @@
-import SafariDetail from "./Safari";
-import Footer from "../Footer";
-import { maraImages } from "../../data/safari";
-import { maraNrbSamburuItenary, maraNrbSamburPromoAd } from "../../data/itenaries";
+import SafariDetail from "../../modals/Safari";
+import Footer from "../../Footer";
+import { maraImages } from "../../../data/safari";
+import { maraNrbSamburPromoAd,maraNrbSamburuItenary } from "./data";
+
 const NairobiSamburuMaraTrip = () => (
   <>
     <SafariDetail
