+++ src/components/kenya-safari/nrb-samburu-mara/data.js
@@ -1,0 +1,54 @@
+export const maraPromoAd = {
+  promoTitle: `Capture the Essence of Africa's Wild Beauty!`,
+  promoBody: `Immerse yourself in the untamed realm of the Masai Mara,
+  home to incredible wildlife spectacles and awe-inspiring landscapes.`,
+  callToAction: `Contact us now to secure your place on an unforgettable safari experience!`
+};
+
+export const maraNrbSamburuItenary = [
+  {
+    day: `Day 1`,
+    title: ' Welcome to Nairobi',
+    activities: [
+      `Arrive in Nairobi, Kenya's vibrant capital. `,
+      `Immerse yourself in the city's culture, visit the Giraffe Centre, or Nairobi National Museum. `,
+      `Retire to a comfortable city hotel, preparing for the wild adventure ahead.`,
+    ]
+  },
+  {
+    day: `Day 2 - 3`,
+    title: `Samburu's Enchanting Wilderness`,
+    activities: [
+      `Journey to Samburu, a realm of unique wildlife.`,
+      `Engage in thrilling game drives to spot the "Samburu Special Five" and explore the Ewaso Ng'iro River's banks.`,
+      `Experience the rich culture of the Samburu people and find serenity in your lodge amidst nature's beauty.`,
+    ]
+  },
+  {
+    day: `Day 4 - 6`,
+    title: `Majestic Masai Mara Awaits`,
+    activities: [
+      `Travel to the renowned Masai Mara, where three days of awe-inspiring adventure await.`,
+      `Witness the Great Migration, Big Five, and countless species on exhilarating game drives.`,
+      `Delve into the Maasai culture with village visits and savor the magic of the Mara under starry African skies.`
+    ]
+  },
+  {
+    day: `Day 7`,
+    title: `Reflecting on an Incredible Journey`,
+    activities: [
+      `Embark on a final game drive in Masai Mara, capturing lingering moments with the wildlife.`,
+      `Return to Nairobi, where you can reminisce about the past week's adventures and prepare for your onward journey.`,
+    ]
+  },
+]
+
+export const maraNrbSamburPromoAd = {
+  promoTitle: `Capture the Essence of Kenya's Beauty!`,
+  promoBody: `Embark on an all-encompassing journey through Kenya's iconic landscapes, 
+  wildlife, and cultures.`,
+  price: `Ksh.15,000`,
+  callToAction: `Contact us now to secure your spot on 
+  a once-in-a-lifetime safari experience 
+  from Nairobi to Samburu and the majestic Masai Mara!`
+};
