+++ src/components/malaysia/data.js
@@ -21,3 +21,63 @@ export const malaysiaImages = [
     alt: "Malaysia"
   }
 ]
+
+// Malysia
+export const malaysiaItenary = [
+  {
+    day: `Day 1`,
+    title: `Arrival in Kuala Lumpur, Malaysia`,
+    activities: [
+      `Welcome to the dynamic city of Kuala Lumpur! Arrive and be transferred to your comfortable accommodation. Take the evening to relax and soak in the vibrant city atmosphere.`
+    ]
+  },
+  {
+    day: `Day 2`,
+    title: `Kuala Lumpur Highlights`,
+    activities: [
+      `Embark on a city tour, exploring Kuala Lumpur's iconic landmarks. Visit the Petronas Twin Towers, the vibrant Central Market, and the historical Merdeka Square. Immerse yourself in the city's rich culture and diversity.`
+    ]
+  },
+  {
+    day: `Day 3`,
+    title: `Batu Caves & Genting Highlands`,
+    activities: [
+      `Explore the stunning Batu Caves, a Hindu temple complex nestled within limestone caves. Continue to Genting Highlands, a mountain resort offering entertainment, casinos, and breathtaking views.`
+    ]
+  },
+  {
+    day: `Day 4`,
+    title: `Transfer to Singapore`,
+    activities: [
+      `Travel to the modern metropolis of Singapore. Check in to your accommodation and spend the afternoon exploring the vibrant streets, shopping districts, and renowned food markets.`
+    ]
+  },
+  {
+    day: `Day 5`,
+    title: `Singapore City Discovery`,
+    activities: [
+      `Delve into Singapore's charm with a city tour. Visit iconic sites like Marina Bay Sands, Gardens by the Bay, and Singapore Botanic Gardens. In the evening, take a memorable stroll along Clarke Quay.`
+    ]
+  },
+  {
+    day: `Day 6`,
+    title: `Sentosa Island Adventure`,
+    activities: [
+      `Embark on a day of excitement at Sentosa Island. Visit Universal Studios for thrilling rides and entertainment. Discover the S.E.A. Aquarium, or simply relax on the island's pristine beaches.`
+    ]
+  },
+  {
+    day: `Day 7`,
+    title: `Departure`,
+    activities: [
+      `Savor your last moments in Singapore with some leisure time for shopping or exploring. Depart with unforgettable memories of your Malaysian and Singaporean adventure.`
+    ]
+  },
+]
+
+export const malaysiaPromoAd = {
+  promoTitle: `Book Your Malaysian & Singaporean Adventure Today! `,
+  promoBody: `Discover the blend of traditions and modernity in Kuala Lumpur and Singapore.`,
+  price: `$2310pp`,
+  callToAction: `Contact us now to secure your spot on a journey that will immerse you in the wonders of Malaysia and Singapore!.`
+};
