+++ src/components/mauritius/data.js
@@ -22,3 +22,48 @@ export const mauritiusImages = [
     alt: "Malaysia"
   }
 ]
+
+export const mauritiusItenary = [
+  {
+    day: `Day 1`,
+    title: `Arrival in Mauritius`,
+    activities: [
+      `Welcome to the stunning island of Mauritius! Upon arrival, you'll be greeted and transferred to your exquisite accommodation. Feel the warm ocean breeze and let the relaxation begin.`
+    ]
+  },
+  {
+    day: `Day 2`,
+    title: `Beach Bliss & Water Adventures`,
+    activities: [
+      `Spend your day basking in the beauty of Mauritius' beaches. Relax under the sun, take a dip in the crystal-clear waters, or engage in exciting water sports. Your seaside retreat promises tranquility and fun.`
+    ]
+  },
+  {
+    day: `Day 3`,
+    title: `Island Exploration`,
+    activities: [
+      `Embark on an island tour to discover the enchanting landscapes of Mauritius. Visit landmarks like the Chamarel Seven Colored Earths, Grand Bassin, and the mesmerizing Chamarel Waterfall.`
+    ]
+  },
+  {
+    day: `Day 4`,
+    title: `Leisure & Culture`,
+    activities: [
+      `Indulge in a day at your leisure. Choose to unwind at the resort's spa, lounge by the pool, or explore local markets for souvenirs. In the evening, savor the island's cuisine and immerse yourself in its vibrant culture.`
+    ]
+  },
+  {
+    day: `Day 5`,
+    title: `Departure`,
+    activities: [
+      `Relish your last moments on the island. Depending on your flight time, enjoy a final stroll on the beach or take a refreshing swim. Depart Mauritius with unforgettable memories of your tropical retreat.`
+    ]
+  },
+]
+
+export const mauritiusPromoAd = {
+  promoTitle: `Book Your Mauritius Retreat Today!`,
+  promoBody: `Unveil the allure of Mauritius' beaches, landscapes, and vibrant culture.`,
+  price: `$1700pp`,
+  callToAction: ` Contact us now to secure your spot on a journey that will immerse you in the blissful tranquility of Mauritius' tropical paradise!`
+};
