+++ src/components/pari-barca-italy/data.js
@@ -31,3 +31,62 @@ export const barcaItalyParisImages = [
     alt: "Barcelona, Italy, France"
   }
 ]
+
+export const parisBarcaItalyItenary = [
+  {
+    day: `Day 1`,
+    title: `Arrival in Paris, France`,
+    activities: [
+      `Begin your European adventure in the romantic city of Paris. Upon arrival, our team will warmly welcome you and transfer you to your charming accommodation. Take in the enchanting atmosphere and get ready for the days of exploration ahead.`
+    ]
+  },
+  {
+    day: `Day 2`,
+    title: `Parisian Charm`,
+    activities: [
+      `Explore the iconic landmarks of Paris. Visit the Eiffel Tower, stroll along the Seine River, and immerse yourself in the artistic wonder of the Louvre Museum. Experience the city's captivating blend of history and culture.`,
+    ]
+  },
+  {
+    day: `Day 3`,
+    title: `Travel to Barcelona, Spain`,
+    activities: [
+      `Fly to the vibrant city of Barcelona. Settle into your accommodation and spend the afternoon wandering through the charismatic streets, admiring Gaudi's architectural masterpieces, and savoring tapas.`,
+    ]
+  },
+  {
+    day: `Day 4`,
+    title: `Barcelona's Treasures`,
+    activities: [
+      `Embark on a comprehensive city tour of Barcelona. Visit the Sagrada Familia, Park Güell, and enjoy the breathtaking views from Montjuïc Hill. Feel the pulse of this creative and vibrant city.`
+    ]
+  },
+  {
+    day: `Day 5`,
+    title: `Explore Rome, Italy`,
+    activities: [
+      `Fly to the eternal city of Rome, Italy. Immerse yourself in its rich history as you visit the Colosseum, Roman Forum, and other ancient sites. Settle into your accommodation and savor authentic Italian cuisine.`
+    ]
+  },
+  {
+    day: `Day 6`,
+    title: `Vatican City & More`,
+    activities: [
+      `Experience the spiritual and artistic wonders of Vatican City. Visit St. Peter's Basilica and the Vatican Museums, home to masterpieces like Michelangelo's Sistine Chapel ceiling. Spend the afternoon exploring more of Rome's charm.`
+    ]
+  },
+  {
+    day: `Day 7`,
+    title: `Reflection and Departure`,
+    activities: [
+      `Savor your final moments in Europe, perhaps with some last-minute shopping or a leisurely stroll. Depart with cherished memories of your European journey.`,
+    ]
+  },
+]
+
+export const parisBaracItalyPromoAd = {
+  promoTitle: `Book Your European Getaway Today!`,
+  promoBody: `Unveil the beauty and diversity of Paris, Barcelona, and Rome. Limited slots are available for this exclusive offer.`,
+  price: `$1790pp`,
+  callToAction: `Contact us now to secure your spot on a journey that promises to immerse you in the cultural riches of Europe!.`
+};
