+++ src/components/rwanda-uganda/data.js
@@ -1,0 +1,104 @@
+export const rwandaUgandaItenary = [
+  {
+    day: `Day 1`,
+    title: `Arrival in Kigali, Rwanda`,
+    activities: [
+      `Arrive in Kigali, the capital of Rwanda. 
+      Our representative will greet you and transfer you to your 
+      luxurious accommodation. Take the day to acclimate and 
+      immerse yourself in the local culture.`
+    ]
+  },
+  {
+    day: `Day 2`,
+    title: `Fly-in to Bwindi Impenetrable Forest, Uganda`,
+    activities: [
+      `Embark on a scenic flight to Bwindi Impenetrable Forest in Uganda. 
+      Land amidst lush greenery and check in to your exclusive lodge. 
+      Prepare for your gorilla trekking adventure with a briefing from local guides.`
+    ]
+  },
+  {
+    day: `Day 3`,
+    title: `Gorilla Trekking in Bwindi`,
+    activities: [
+      `Today is the day you've been waiting for. 
+      Trek through the dense forest in search of the 
+      magnificent mountain gorillas. Witness these gentle 
+      giants in their natural habitat, an experience that 
+      will forever leave an imprint on your soul.`
+    ]
+  },
+  {
+    day: `4`,
+    title: `Helicopter Safari to Queen Elizabeth National Park`,
+    activities: [
+      `Soar over Uganda's landscapes in a helicopter as you head to 
+      Queen Elizabeth National Park. Marvel at the vast savannas, 
+      waterholes, and diverse wildlife below. 
+      Check in to your luxury lodge overlooking the park.`
+    ]
+  },
+  {
+    day: `Day 5`,
+    title: `Savanna Game Drive and Boat Safari`,
+    activities: [
+      `Embark on a thrilling game drive in Queen Elizabeth National Park, 
+       where lions, elephants, and buffalo roam. In the afternoon, 
+       set sail on the Kazinga Channel for a boat safari, 
+       encountering hippos, crocodiles, and numerous bird species.`
+    ]
+  },
+  {
+    day: `Day 6`,
+    title: `Return to Rwanda`,
+    activities: [
+      `Fly back to Kigali and visit the Genocide Memorial to pay
+      respects to Rwanda's history. Enjoy a city tour before returning
+       to your accommodation for relaxation.`
+    ]
+  },
+  {
+    day: `Day 7`,
+    title: `Volcanoes National Park - Golden Monkeys`,
+    activities: [
+      `Transfer to Volcanoes National Park and 
+       embark on a trek to meet the playful golden monkeys. 
+       Enjoy the vibrant forest and breathtaking views..`
+    ]
+  },
+  {
+    day: `Day 8`,
+    title: `Gorilla Trekking in Volcanoes National Park`,
+    activities: [
+      `Another day, another gorilla encounter. Trek through Volcanoes National Park, and this time, meet a different gorilla family. Reflect on the incredible privilege of observing these incredible creatures up close.`
+    ]
+  },
+  {
+    day: `Day 9`,
+    title: `Community Interaction and Departure`,
+    activities: [
+      `Engage with the local community, learning about their way of life and conservation efforts. Depart with a deeper understanding of the delicate balance between wildlife and humans.`
+    ]
+  },
+]
+
+export const rwandaUgandaPromoAd = {
+  promoTitle: `Ready for Your Gorilla Adventure?`,
+  promoBody: `Experience the majesty of mountain gorillas 
+  and the diverse beauty of Rwanda and Uganda. 
+  Our meticulously curated itinerary offers helicopter flights, 
+  luxury lodges, thrilling treks, and meaningful interactions. 
+  `,
+  callToAction: `Contact us to embark on a journey that 
+  will redefine your connection with nature and wildlife.`
+};
+
+export const rwandaUgandaExtras = [
+  `Helicopter and Fly-in safaris`,
+  `Luxury and Exclusive safaris`,
+  `Primate and Savanna safaris`,
+  `Group and Family safaris`,
+  `Walking and Trekking safaris`,
+  ` Eco and Community safaris`
+]
