+++ src/components/turkey/data.js
@@ -31,3 +31,48 @@ export const turkeyImages = [
     alt: "Turkey"
   }
 ]
+
+export const turkeyItenary = [
+  {
+    day: `Day 1`,
+    title: `Arrival in Istanbul, Turkey`,
+    activities: [
+      `Welcome to the enchanting city of Istanbul! Upon arrival, our team will greet you and transfer you to your comfortable accommodation. Take in the blend of history and modernity that defines this vibrant city.`,
+    ]
+  },
+  {
+    day: `Day 2`,
+    title: `Explore Historical Istanbul`,
+    activities: [
+      `Embark on a captivating exploration of Istanbul's historical sites. Visit the iconic Hagia Sophia, the magnificent Blue Mosque, and the Grand Bazaar, where you can immerse yourself in the rich tapestry of Turkish culture.`,
+    ]
+  },
+  {
+    day: `Day 3`,
+    title: `Bosphorus Experience`,
+    activities: [
+      `Embark on a cruise along the Bosphorus, offering panoramic views of Istanbul's stunning skyline. Visit the opulent Dolmabahçe Palace and explore the charming neighborhoods that line the waterway.`,
+    ]
+  },
+  {
+    day: `Day 4`,
+    title: `Ottoman Heritage`,
+    activities: [
+      `Step into the opulent past of the Ottoman Empire. Visit Topkapi Palace, once the residence of Ottoman sultans, and explore the intricate details of the palace's architecture and treasures.`
+    ]
+  },
+  {
+    day: `Day 5`,
+    title: `Departure`,
+    activities: [
+      `Savor your final moments in Istanbul, perhaps with a leisurely stroll or some last-minute shopping. Depart with cherished memories of your Turkish journey and the allure of Istanbul's cultural heritage.`,
+    ]
+  },
+]
+
+export const turkeyPromoAd = {
+  promoTitle: `Unveil the excitement of Bangkok's city life and the tranquility of Phuket's beaches.`,
+  promoBody: `Embark on a journey through the heart of Turkey's cultural gem, Istanbul. Starting from just $1660 per person, this offer ensures you experience the city's rich history, vibrant markets, and iconic landmarks.`,
+  price: `$1660pp`,
+  callToAction: `Contact us now to secure your spot on a journey that promises an immersive experience in the heart of Turkish culture.`
+};
