--- src/data/about.js
@@ -1,17 +1,0 @@
-export const leftServices = [
-  `Inbound & Outbound Tours Travel`,
-  `Air Ticketing`,
-  `Air transfer`,
-  `Visa & Travel Insurance Processing`,
-  `Nature Expedition & People Cultural Programme`,
-  `Corporate Team Building`
-];
-
-export const rightServices = [
-  `Budget, Mid-range Safaris`,
-  `Hotel Booking`,
-  `Wildlife Safaris`,
-  `Beach Holidays`,
-  `Day Excursion Trips`,
-  `Honeymoon Packages`
-];
