--- src/data/cities.js
@@ -1,11 +1,0 @@
-import dubai from '../assets/images/cities/dubai.jpg';
-import joburg from '../assets/images/cities/capetown.jpg'
-import thailand from '../assets/images/cities/thailand.jpg';
-import turkey from '../assets/images/cities/turkey.jpg';
-import rome from '../assets/images/cities/rome.jpg';
-import mauritius from '../assets/images/cities/mauritius.jpg';
-import china from '../assets/images/cities/china.jpg';
-import singapore from '../assets/images/cities/singapore.jpg';
-import israel from '../assets/images/cities/israel.jpg';
-
-export const cities = [dubai, joburg, singapore, mauritius, china, israel, rome, thailand, turkey]
