--- src/data/international.js
@@ -1,90 +1,0 @@
-import dubai from '../assets/images/cities/dubai.jpg';
-import joburg from '../assets/images/cities/capetown.jpg'
-import thailand from '../assets/images/cities/thailand.jpg';
-import turkey from '../assets/images/cities/turkey.jpg';
-import rome from '../assets/images/cities/rome.jpg';
-import mauritius from '../assets/images/cities/mauritius.jpg';
-import china from '../assets/images/cities/china.jpg';
-import singapore from '../assets/images/cities/singapore.jpg';
-import israel from '../assets/images/cities/israel.jpg';
-
-export const intlTourDestinations = [
-  {
-    title: "Discover Dubai",
-    image: dubai,
-    days: "5",
-    nights: "4",
-    description: `Our meticulously crafted itinerary ensures you experience the city's highlights. All starting from $650 per person, this offer provides an unbeatable value for an unforgettable experience`,
-    url: "dubai",
-  },
-  {
-    title: "Johannesburg & Sun City Delight",
-    description: `Embark on an unforgettable 7-day journey through the dynamic city of Johannesburg and the luxurious oasis of Sun City. Starting from just $1495 per person.`,
-    image: joburg,
-    days: "7",
-    nights: "6",
-    url: "joburg",
-  },
-  {
-    title: "Malaysia & Singapore Exploration",
-    image: singapore,
-    days: "7",
-    nights: "6",
-    description: `Embark on a captivating 7-day journey through the cultural richness of Kuala Lumpur and the modern allure of Singapore. Starting from just $2310 per person`,
-    url: "malaysia",
-  },
-  {
-    title: "Mauritius Escapade",
-    image: mauritius,
-    days: "4",
-    nights: "3",
-    description: `Experience the paradisiacal beauty of Mauritius in just 5 days! Starting from $1700 per person, this offer ensures you experience the best of the island's beaches, culture, and natural wonders.`,
-    url: "mauritius",
-  },
-  {
-    title: "China Canton Fair Trade Exhibition Adventure",
-    image: china,
-    days: "10",
-    nights: "9",
-    description: `Starting from $1890 per person, you ge to engage in the dynamic world of trade at the Canton Fair Exhibition as well as an opportunity to explore Chinese culture.`,
-    url: `china`,
-  },
-  {
-    title: "Sacred Pilgrimage Tour: Egypt & Israel Journey of Faith",
-    image: israel,
-    days: "9",
-    nights: "8",
-    description: `Embark on a spiritual pilgrimage through Egypt and Israel, exploring sacred sites and connecting with faith. Starting from just $2450 per person.`,
-    url: "egypt-israel"
-  },
-  {
-    title: "Paris, Barcelona & Italy Adventure",
-    image: rome,
-    days: "7",
-    nights: "6",
-    description: `Experience the allure of three European gems in just 7 days! Starting from $1790 per person, this offer provides you with a taste of Parisian fashion, Barcelona's vibrancy, and the historical wonders of Rome.`,
-    url: "italy-barca-paris"
-  },
-  {
-    title: "Bangkok & Phuket Getaway: 7 Days & 6 Nights of Thai Adventure!",
-    image: thailand,
-    days: "7",
-    nights: "6",
-    description: `Get to Experience the best of both worlds in 
-    Thailand's bustling capital and tropical paradise for just $1875 per person. 
-    Experience the beauty, culture, 
-    and relaxation Thailand has to offer.`,
-    url: "bangkok"
-  },
-  {
-    title: "Istanbul Discovery: 5 Days & 4 Nights Journey through Turkey's Cultural Heart",
-    image: turkey,
-    days: "5",
-    nights: "4",
-    description: `Embark on a journey through the heart of Turkey's cultural gem, 
-    Istanbul. Starting from just $1660 per person, you experience the city's rich history, 
-    vibrant markets, and iconic landmarks.`,
-    url: "turkey"
-  },
-]
-
