--- src/data/itenaries.js
@@ -1,1002 +1,0 @@
-export const masaaiMaraIternary = [
-  {
-    day: `Day 1`,
-    title: 'Journey to Masai Mara',
-    activities: [
-      `Embark on a thrilling adventure as we transport you to the iconic Masai Mara.`,
-      `Feel the anticipation build as you approach this legendary wilderness.`,
-      `Settle into your luxurious lodge, surrounded by the sights and sounds of the African savanna.`
-    ],
-  },
-  {
-    day: `Day 2`,
-    title: `Unveiling the Mara's Wonders`,
-    activities: [
-      `Rise with the sun for a day of unparalleled discovery.
-       Engage in heart-pounding game drives as the Mara's wildlife awakens
-       Witness lions on the prowl, elephants at play, and graceful giraffes on the horizon.
-       Experience the untamed beauty that has captivated explorers for generations.`,
-    ],
-  },
-
-  {
-    day: `Day 3`,
-    title: 'Farewell to the Mara',
-    activities: [
-      `Capture final glimpses of the Mara's wildlife on a morning game drive, etching these memories in your heart,
-      As the sun sets on your journey, return to Nairobi with a profound connection to the wild,
-      Your adventure concludes with drop-off at your preferred location.`,
-    ],
-  },
-];
-
-export const masaaiMaraPromoAd = {
-  promoTitle: `Capture the Essence of Africa's Wilderness!`,
-  promoBody: `Experience the awe-inspiring landscapes and incredible wildlife 
-  of Masai Mara in just 3 days. `,
-  callToAction: `
-  Contact us now to secure your spot on a journey 
-  that promises unforgettable 
-  moments in the heart of the wild!`
-}
-
-export const lakeNakuruIternary = [
-  {
-    day: `Day 1`,
-    title: 'Lake Nakuru Welcomes You',
-    activities: [
-      `Depart from Nairobi and journey to Lake Nakuru`,
-      `Marvel at the flamingo spectacle and diverse birdlife.`,
-      `Thrilling game drive to spot rhinos, lions, and more.`,
-      `Overnight stay in a luxurious lodge by the lake.`
-    ]
-  },
-  {
-    day: `Day 2`,
-    title: `Into the Heart of Masai Mara`,
-    activities: [
-      `Travel to the enchanting Masai Mara.`,
-      `Afternoon game drive in search of the Big Five`,
-      `Immerse yourself in the untamed beauty of the Mara.`,
-      `Unwind at our carefully selected lodge.`,
-    ]
-  },
-  {
-    day: `Day 3`,
-    title: 'The Great Migration Experience',
-    activities: [
-      `Embark on a full-day game drive to witness the Great Migration.`,
-      `Capture unforgettable moments as herds traverse the plains.`,
-      `Encounter predators and their thrilling pursuits.`,
-      `Evening relaxation and overnight stay in Masai Mara.`,
-    ]
-  },
-  {
-    day: `Day 4`,
-    title: 'Farewell to the Wilderness',
-    activities: [
-      `Capture final glimpses of Mara's wildlife on a morning game drive.`,
-      `Return to Nairobi, cherishing memories of a lifetime.`,
-      `Your adventure concludes with drop-off at your preferred location.`,
-    ],
-  },
-];
-
-export const lakeNakPromoAd = {
-  promoTitle: `Book Now for an Unforgettable Safari!`,
-  promoBody: `Don't miss this chance to explore two of Kenya's most iconic destinations. 
-  Immerse yourself in the beauty of Lake Nakuru and witness 
-  the grandeur of the Masai Mara. Limited slots available.`,
-  callToAction: `Contact us to secure your spot on 
-  this extraordinary 4-day journey into the wild!
-  `
-};
-
-
-export const maraItenary = [
-  {
-    day: `Day 1`,
-    title: 'Into the Heart of the Mara',
-    activities: [
-      `Embark on an exhilarating journey as we transport you to the captivating Masai Mara.`,
-      `Brace yourself for an afternoon game drive,
-       delving into the heart of this iconic wilderness.`,
-      `Encounter elephants, lions, and the diverse wildlife that calls Mara home.`,
-      `Retreat to our carefully selected lodge and relish the tranquility of the African night.`
-    ]
-  },
-  {
-    day: `Day 2`,
-    title: 'The Great Mara Safari',
-    activities: [
-      `Rise with the sun for a full-day adventure that unfolds the magic of the Masai Mara.`,
-      `Witness the majestic landscape transform as herds of wildebeest and zebras roam the plains in the epic Great Migration.`,
-      `Capture nature's drama on your camera as predators roam in pursuit.`,
-      `As dusk falls, unwind at our lodge while the sounds of the wild serenade you.`
-    ]
-  },
-  {
-    day: `Day 3`,
-    title: `Farewell to Mara's Splendors`,
-    activities: [
-      `Your morning game drive holds the promise of last-minute wildlife encounters`,
-      `Bid farewell to the Mara as you savor these final moments in the wilderness.`,
-      `Return to Nairobi with memories etched in your heart and camera roll, arriving in the late afternoon.`
-    ]
-  },
-]
-
-export const maraPromoAd = {
-  promoTitle: `Capture the Essence of Africa's Wild Beauty!`,
-  promoBody: `Immerse yourself in the untamed realm of the Masai Mara,
-  home to incredible wildlife spectacles and awe-inspiring landscapes.`,
-  callToAction: `Contact us now to secure your place on an unforgettable safari experience!`
-};
-
-export const maraNrbSamburuItenary = [
-  {
-    day: `Day 1`,
-    title: ' Welcome to Nairobi',
-    activities: [
-      `Arrive in Nairobi, Kenya's vibrant capital. `,
-      `Immerse yourself in the city's culture, visit the Giraffe Centre, or Nairobi National Museum. `,
-      `Retire to a comfortable city hotel, preparing for the wild adventure ahead.`,
-    ]
-  },
-  {
-    day: `Day 2 - 3`,
-    title: `Samburu's Enchanting Wilderness`,
-    activities: [
-      `Journey to Samburu, a realm of unique wildlife.`,
-      `Engage in thrilling game drives to spot the "Samburu Special Five" and explore the Ewaso Ng'iro River's banks.`,
-      `Experience the rich culture of the Samburu people and find serenity in your lodge amidst nature's beauty.`,
-    ]
-  },
-  {
-    day: `Day 4 - 6`,
-    title: `Majestic Masai Mara Awaits`,
-    activities: [
-      `Travel to the renowned Masai Mara, where three days of awe-inspiring adventure await.`,
-      `Witness the Great Migration, Big Five, and countless species on exhilarating game drives.`,
-      `Delve into the Maasai culture with village visits and savor the magic of the Mara under starry African skies.`
-    ]
-  },
-  {
-    day: `Day 7`,
-    title: `Reflecting on an Incredible Journey`,
-    activities: [
-      `Embark on a final game drive in Masai Mara, capturing lingering moments with the wildlife.`,
-      `Return to Nairobi, where you can reminisce about the past week's adventures and prepare for your onward journey.`,
-    ]
-  },
-]
-
-export const maraNrbSamburPromoAd = {
-  promoTitle: `Capture the Essence of Kenya's Beauty!`,
-  promoBody: `Embark on an all-encompassing journey through Kenya's iconic landscapes, 
-  wildlife, and cultures.`,
-  callToAction: `Contact us now to secure your spot on 
-  a once-in-a-lifetime safari experience 
-  from Nairobi to Samburu and the majestic Masai Mara!`
-};
-
-export const amboseliTsavoItenary = [
-  {
-    day: `Day 1`,
-    title: `Journey to Amboseli's Majesty`,
-    activities: [
-      `Commence your adventure as we whisk you away to Amboseli National Park.`,
-      `et against the backdrop of Mount Kilimanjaro, experience breathtaking views and remarkable elephant herds.`,
-      `Engage in an afternoon game drive, capturing the essence of Amboseli's unique charm.`,
-    ]
-  },
-  {
-    day: `Day 2`,
-    title: `Amboseli's Enchanting Wildlife`,
-    activities: [
-      `Spend a full day exploring Amboseli's diverse landscapes and its abundant wildlife.`,
-      `Lions, cheetahs, and a plethora of bird species await your discovery`,
-      `Capture the iconic elephants with Kilimanjaro's snow-capped peak as a backdrop.`,
-    ]
-  },
-  {
-    day: `Day 3`,
-    title: `Tsavo West's Hidden Gems`,
-    activities: [
-      `Embark on a scenic journey to Tsavo West National Park, a realm of hidden treasures.`,
-      `Venture into the Mzima Springs to observe hippos and crocodiles underwater.`,
-      `Explore the Shetani Lava Flows and marvel at the ancient landscapes of this unique wilderness`
-    ]
-  },
-  {
-    day: `Day 4`,
-    title: `Reflecting on Nature's Grandeur`,
-    activities: [
-      `Conclude your journey with a morning game drive in Tsavo West`,
-      `Relish the final moments with the wildlife and landscapes that have left an indelible mark on your heart`,
-      `Return to your starting point with memories of Kenya's untamed beauty.`
-    ]
-  },
-]
-
-export const amboseliTsavoPromoAd = {
-  promoTitle: `Capture the Essence of Kenya's Wilderness!`,
-  promoBody: `Embark on a captivating 4-day journey through Amboseli 
-  and Tsavo West's natural wonders. `,
-  callToAction: `Contact us now to secure your spot 
-  on an unforgettable safari experience that unveils the 
-  beauty of Kenya's landscapes and wildlife!`
-};
-
-export const nrbAmboseliMaraItenary = [
-  {
-    day: `Day 1`,
-    title: 'Arrival in Nairobi',
-    activities: [
-      `Your journey begins in Nairobi, Kenya's dynamic capital.`,
-      `Explore its vibrant culture, visit the David Sheldrick Wildlife Trust, or indulge in local markets`,
-      `Settle into your city accommodation and prepare for the diverse adventure ahead.`,
-    ]
-  },
-  {
-    day: `Day 2 - 3`,
-    title: `Amboseli's Majesty`,
-    activities: [
-      `Venture to Amboseli National Park, an iconic realm beneath the gaze of Mount Kilimanjaro.`,
-      `Immerse yourself in unrivaled landscapes and experience thrilling game drives.`,
-      `Encounter herds of elephants against the backdrop of Africa's tallest peak.`,
-    ]
-  },
-  {
-    day: `Day 4 - 5`,
-    title: 'The Enchanting Masai Mara',
-    activities: [
-      `Travel to the world-famous Masai Mara, where two days of unforgettable experiences await. `,
-      `Witness the grandeur of the Great Migration and spot the Big Five on exhilarating game drives`,
-      `Engage with Maasai communities and soak in the magic of the Mara's sunsets.`,
-    ]
-  },
-  {
-    day: `Day 6`,
-    title: 'Reflecting on an Extraordinary Journey',
-    activities: [
-      `Conclude your adventure with a morning game drive in Masai Mara, 
-      capturing last memories of the wildlife.`,
-      `Return to Nairobi, cherishing the encounters and landscapes that have left an indelible mark on your heart.`,
-    ],
-  },
-
-];
-
-export const nrbAmoseliMaraPromo = {
-  promoTitle: `Capture the Essence of Kenya's Diversity! `,
-  promoBody: `Embark on an all-encompassing 6-day expedition, 
-  spanning city life, breathtaking landscapes, 
-  and remarkable wildlife encounters.`,
-  callToAction: `Contact us now to secure your place on an unforgettable 
-  safari adventure from Nairobi to Amboseli and the enchanting Masai Mara!`
-};
-
-
-export const nrbAmboseliMtKenyaItenary = [
-  {
-    day: `Day 1`,
-    title: `Welcome to Nairobi`,
-    activities: [
-      `Arrive in Nairobi, Kenya's bustling capital.`,
-      `Explore its vibrant markets, museums, and wildlife centers.`,
-      `Settle into your city accommodation and prepare for the extraordinary journey ahead.`,
-    ]
-  },
-  {
-    day: `Day 2 - 3`,
-    title: `Amboseli's Majestic Charm`,
-    activities: [
-      `Head to Amboseli National Park, where the mighty Mount Kilimanjaro serves as the backdrop.`,
-      `Engage in exhilarating game drives to witness elephants, lions, and a diversity of species.`,
-      `Relish the magic of the African wilderness.`,
-    ]
-  },
-  {
-    day: `Day 4`,
-    title: `Exploring the Mystical Mt. Kenya`,
-    activities: [
-      `Travel to the awe-inspiring Mount Kenya region.`,
-      `Immerse yourself in the lush landscapes, witness diverse flora and fauna, and take in the serenity of the highland environment.`,
-      `Overnight stay near the majestic mountain.`,
-    ]
-  },
-  {
-    day: `Day 5`,
-    title: `Lake Nakuru's Pink Spectacle`,
-    activities: [
-      `Journey to Lake Nakuru National Park, a paradise for bird enthusiasts`,
-      `Witness the breathtaking sight of flamingos painting the lake pink.`,
-      `Encounter rhinos, lions, and other wildlife amidst stunning vistas.`
-    ]
-  },
-  {
-    day: `Day 6-7`,
-    title: `Majestic Masai Mara Awaits`,
-    activities: [
-      `Travel to the iconic Masai Mara for two days of unparalleled adventure.`,
-      `Witness the Great Migration, spot the Big Five, and capture unforgettable moments on game drives.`,
-      `Engage with Maasai communities and savor the African sunset.`
-    ]
-  },
-  {
-    day: 'Day 8',
-    title: `Reflections on an Epic Journey`,
-    activities: [
-      `Conclude your safari with a final game drive in Masai Mara.`,
-      `Return to Nairobi, carrying memories of diverse landscapes, 
-      captivating wildlife, and rich cultural encounters`,
-    ]
-
-  },
-]
-
-export const nrbAmboseliMtKenyaPromoAd = {
-  promoTitle: `Capture the Essence of Kenya's Diversity and Splendor!`,
-  promoBody: `
-  Embark on the journey of a lifetime, spanning Kenya's iconic 
-  landscapes, vibrant cultures, and mesmerizing wildlife encounters.`,
-  callToAction: `Contact us now to secure your place on an unforgettable 
-  safari that explores Nairobi, Amboseli, Mt. Kenya, Lake Nakuru, 
-  and the captivating Masai Mara!`
-};
-
-export const nrbAmboseliLewaItenary = [
-  {
-    day: `Day 1`,
-    title: `Arrival in Nairobi`,
-    activities: [
-      `Begin your adventure in Nairobi, Kenya's dynamic capital.`,
-      `Immerse yourself in its urban culture, explore local markets, or visit wildlife sanctuaries.`,
-      `Settle into your city accommodation and anticipate the enchanting journey ahead.`,
-    ]
-  },
-  {
-    day: `Day 2 - 3`,
-    title: `Amboseli's Grace and Majesty`,
-    activities: [
-      `Embark to Amboseli National Park, where the magnificence of Mount Kilimanjaro unfolds.`,
-      `Engage in exhilarating game drives to witness elephants, lions, and the park's iconic landscapes.`,
-      `Capture unforgettable moments under Africa's tallest peak.`,
-    ]
-  },
-  {
-    day: `Day 4 - 5`,
-    title: `Lewa Conservancy's Wilderness`,
-    activities: [
-      `Journey to the Lewa Conservancy, a private haven dedicated to conservation.`,
-      `Embark on game drives, encountering rare species such as Grevy's zebras and black rhinos.`,
-      `Engage in conservation activities and experience the untamed beauty of this exclusive sanctuary.`,
-    ]
-  },
-  {
-    day: `Day 6 - 7`,
-    title: `Majestic Masai Mara Awaits`,
-    activities: [
-      `Travel to the renowned Masai Mara for two days of remarkable adventure.`,
-      `Witness the Great Migration, spot the Big Five, and capture the essence of the African wilderness on exhilarating game drives.`,
-      `Engage with Maasai communities and relish the Mara's captivating sunsets.`
-    ]
-  },
-  {
-    day: `Day 8`,
-    title: `Reflections on a Multifaceted Journey`,
-    activities: [
-      `Conclude your safari with a final game drive in Masai Mara.`,
-      ` Return to Nairobi, carrying cherished memories of diverse landscapes, captivating wildlife, and rich cultural encounters.`,
-    ]
-  },
-]
-
-export const nrbAmboseliLewaPromo = {
-  promoTitle: `Capture the Essence of Kenya's Diversity and Conservation Efforts!`,
-  promoBody: `
-  Embark on an extraordinary 8-day journey, 
-  traversing Kenya's diverse landscapes, 
-  rich cultures, and incredible wildlife encounters.`,
-  callToAction: `Contact us now to secure your spot on a once-in-a-lifetime exploration 
-  that encompasses Nairobi, Amboseli, 
-  Lewa Conservancy, and the captivating Masai Mara! `
-};
-
-export const rwandaUgandaItenary = [
-  {
-    day: `Day 1`,
-    title: `Arrival in Kigali, Rwanda`,
-    activities: [
-      `Arrive in Kigali, the capital of Rwanda. 
-      Our representative will greet you and transfer you to your 
-      luxurious accommodation. Take the day to acclimate and 
-      immerse yourself in the local culture.`
-    ]
-  },
-  {
-    day: `Day 2`,
-    title: `Fly-in to Bwindi Impenetrable Forest, Uganda`,
-    activities: [
-      `Embark on a scenic flight to Bwindi Impenetrable Forest in Uganda. 
-      Land amidst lush greenery and check in to your exclusive lodge. 
-      Prepare for your gorilla trekking adventure with a briefing from local guides.`
-    ]
-  },
-  {
-    day: `Day 3`,
-    title: `Gorilla Trekking in Bwindi`,
-    activities: [
-      `Today is the day you've been waiting for. 
-      Trek through the dense forest in search of the 
-      magnificent mountain gorillas. Witness these gentle 
-      giants in their natural habitat, an experience that 
-      will forever leave an imprint on your soul.`
-    ]
-  },
-  {
-    day: `4`,
-    title: `Helicopter Safari to Queen Elizabeth National Park`,
-    activities: [
-      `Soar over Uganda's landscapes in a helicopter as you head to 
-      Queen Elizabeth National Park. Marvel at the vast savannas, 
-      waterholes, and diverse wildlife below. 
-      Check in to your luxury lodge overlooking the park.`
-    ]
-  },
-  {
-    day: `Day 5`,
-    title: `Savanna Game Drive and Boat Safari`,
-    activities: [
-      `Embark on a thrilling game drive in Queen Elizabeth National Park, 
-       where lions, elephants, and buffalo roam. In the afternoon, 
-       set sail on the Kazinga Channel for a boat safari, 
-       encountering hippos, crocodiles, and numerous bird species.`
-    ]
-  },
-  {
-    day: `Day 6`,
-    title: `Return to Rwanda`,
-    activities: [
-      `Fly back to Kigali and visit the Genocide Memorial to pay
-      respects to Rwanda's history. Enjoy a city tour before returning
-       to your accommodation for relaxation.`
-    ]
-  },
-  {
-    day: `Day 7`,
-    title: `Volcanoes National Park - Golden Monkeys`,
-    activities: [
-      `Transfer to Volcanoes National Park and 
-       embark on a trek to meet the playful golden monkeys. 
-       Enjoy the vibrant forest and breathtaking views..`
-    ]
-  },
-  {
-    day: `Day 8`,
-    title: `Gorilla Trekking in Volcanoes National Park`,
-    activities: [
-      `Another day, another gorilla encounter. Trek through Volcanoes National Park, and this time, meet a different gorilla family. Reflect on the incredible privilege of observing these incredible creatures up close.`
-    ]
-  },
-  {
-    day: `Day 9`,
-    title: `Community Interaction and Departure`,
-    activities: [
-      `Engage with the local community, learning about their way of life and conservation efforts. Depart with a deeper understanding of the delicate balance between wildlife and humans.`
-    ]
-  },
-]
-
-export const rwandaUgandaPromoAd = {
-  promoTitle: `Ready for Your Gorilla Adventure?`,
-  promoBody: `Experience the majesty of mountain gorillas 
-  and the diverse beauty of Rwanda and Uganda. 
-  Our meticulously curated itinerary offers helicopter flights, 
-  luxury lodges, thrilling treks, and meaningful interactions. 
-  `,
-  callToAction: `Contact us to embark on a journey that 
-  will redefine your connection with nature and wildlife.`
-};
-
-export const rwandaUgandaExtras = [
-  `Helicopter and Fly-in safaris`,
-  `Luxury and Exclusive safaris`,
-  `Primate and Savanna safaris`,
-  `Group and Family safaris`,
-  `Walking and Trekking safaris`,
-  ` Eco and Community safaris`
-]
-
-// Dubai
-export const dubai4NightsItenary = [
-  {
-    day: `Day 1`,
-    title: `Arrival in Dubai`,
-    activities: [
-      `Arrive in the dazzling city of Dubai, where modern marvels meet rich tradition. Our representative will greet you and transfer you to your luxurious accommodation. Take the evening to unwind and marvel at the city's iconic skyline.`
-    ]
-  },
-  {
-    day: `Day 2`,
-    title: `City Exploration and Desert Safari`,
-    activities: [
-      `Embark on a guided city tour, visiting landmarks like the Burj Khalifa, Dubai Mall, and the Palm Jumeirah. In the afternoon, get ready for an exhilarating desert safari. Dune bashing, camel rides, and a stunning desert sunset await.`
-    ]
-  },
-  {
-    day: `Day 3`,
-    title: `Luxury and Shopping Extravaganza`,
-    activities: [
-      `Indulge in a day of shopping and relaxation. Visit the renowned Dubai Mall for high-end shopping and experience the thrill of the indoor Dubai Aquarium. Unwind in your hotel's spa or take a leisurely stroll along Jumeirah Beach.`
-    ]
-  },
-  {
-    day: `Day 4`,
-    title: `Abu Dhabi Excursion`,
-    activities: [
-      `Take a day trip to Abu Dhabi, the UAE's capital. Explore the majestic Sheikh Zayed Grand Mosque, the stunning Emirates Palace, and the futuristic Louvre Abu Dhabi. Return to Dubai in the evening, where the city's lights will dazzle you.`
-    ]
-  },
-  {
-    day: `Day 5`,
-    title: `Farewell to Dubai`,
-    activities: [
-      `Before departure, visit the historic Dubai Creek and engage in a traditional abra boat ride. Capture a few last-minute memories before bidding farewell to this extraordinary city.`
-    ]
-  },
-]
-
-export const dubai4NightsPromoAd = {
-  promoTitle: `Book Your Dubai Getaway Today!`,
-  promoBody: `Experience the magic of Dubai's skyscrapers, deserts, and luxury. Limited slots are available for this exclusive offer.`,
-  callToAction: `Contact us to embark on a journey that 
-  will redefine your connection with nature and wildlife.`
-};
-
-
-// Joburg 
-export const joburgItenary = [
-  {
-    day: `Day 1`,
-    title: `Arrival in Johannesburg`,
-    activities: [
-      `Arrive in the vibrant city of Johannesburg, your gateway to South African wonders. Our warm welcome awaits as we transfer you to your comfortable accommodation. Spend the evening relaxing and anticipating the exciting days ahead.`
-    ]
-  },
-  {
-    day: `Day 2`,
-    title: `Johannesburg Exploration`,
-    activities: [
-      `Embark on a city tour, delving into Johannesburg's rich history and culture. Visit the Apartheid Museum, Constitution Hill, and explore the vibrant neighborhoods that reflect the city's diversity.`
-    ]
-  },
-  {
-    day: `Day 3`,
-    title: `Sun City Arrival`,
-    activities: [
-      `Travel to the luxurious Sun City resort, an oasis of entertainment and relaxation. Check in to your resort and immerse yourself in the lush surroundings, pristine pools, and exhilarating activities.`
-    ]
-  },
-  {
-    day: `Day 4`,
-    title: `Sun City Adventures`,
-    activities: [
-      `Enjoy a day of endless possibilities in Sun City. From thrilling water activities at the Valley of Waves to exploring the exquisite gardens and entertainment complexes, there's something for everyone.`
-    ]
-  },
-  {
-    day: `Day 5`,
-    title: `Pilanesberg National Park Safari`,
-    activities: [
-      `Embark on a safari adventure in Pilanesberg National Park, adjacent to Sun City. Encounter Africa's iconic wildlife, including the Big Five, on a thrilling game drive.`
-    ]
-  },
-  {
-    day: `Day 6`,
-    title: `Return to Johannesburg`,
-    activities: [
-      `Bid farewell to Sun City and return to Johannesburg. Spend the afternoon at leisure, perhaps indulging in some last-minute shopping or exploring more of the city's attractions.`
-    ]
-  },
-  {
-    day: `Day 7`,
-    title: `Departure`,
-    activities: [
-      `Before departure, take a moment to reflect on your South African adventure. Depending on your flight time, you may have the opportunity for a final exploration or relaxation. Depart with cherished memories and the promise of returning.`
-    ]
-  },
-]
-
-export const joburgPromoAd = {
-  promoTitle: `Book Your South African Adventure Today!`,
-  promoBody: `Discover the vibrant city life of Johannesburg and the leisurely paradise of Sun City. Limited slots are available for this exclusive offer. `,
-  callToAction: `Contact us now to secure your spot on a journey that will immerse you in the splendor and diversity of South Africa's attractions.`
-};
-
-// Malysia
-export const malaysiaItenary = [
-  {
-    day: `Day 1`,
-    title: `Arrival in Kuala Lumpur, Malaysia`,
-    activities: [
-      `Welcome to the dynamic city of Kuala Lumpur! Arrive and be transferred to your comfortable accommodation. Take the evening to relax and soak in the vibrant city atmosphere.`
-    ]
-  },
-  {
-    day: `Day 2`,
-    title: `Kuala Lumpur Highlights`,
-    activities: [
-      `Embark on a city tour, exploring Kuala Lumpur's iconic landmarks. Visit the Petronas Twin Towers, the vibrant Central Market, and the historical Merdeka Square. Immerse yourself in the city's rich culture and diversity.`
-    ]
-  },
-  {
-    day: `Day 3`,
-    title: `Batu Caves & Genting Highlands`,
-    activities: [
-      `Explore the stunning Batu Caves, a Hindu temple complex nestled within limestone caves. Continue to Genting Highlands, a mountain resort offering entertainment, casinos, and breathtaking views.`
-    ]
-  },
-  {
-    day: `Day 4`,
-    title: `Transfer to Singapore`,
-    activities: [
-      `Travel to the modern metropolis of Singapore. Check in to your accommodation and spend the afternoon exploring the vibrant streets, shopping districts, and renowned food markets.`
-    ]
-  },
-  {
-    day: `Day 5`,
-    title: `Singapore City Discovery`,
-    activities: [
-      `Delve into Singapore's charm with a city tour. Visit iconic sites like Marina Bay Sands, Gardens by the Bay, and Singapore Botanic Gardens. In the evening, take a memorable stroll along Clarke Quay.`
-    ]
-  },
-  {
-    day: `Day 6`,
-    title: `Sentosa Island Adventure`,
-    activities: [
-      `Embark on a day of excitement at Sentosa Island. Visit Universal Studios for thrilling rides and entertainment. Discover the S.E.A. Aquarium, or simply relax on the island's pristine beaches.`
-    ]
-  },
-  {
-    day: `Day 7`,
-    title: `Departure`,
-    activities: [
-      `Savor your last moments in Singapore with some leisure time for shopping or exploring. Depart with unforgettable memories of your Malaysian and Singaporean adventure.`
-    ]
-  },
-]
-
-export const malaysiaPromoAd = {
-  promoTitle: `Book Your Malaysian & Singaporean Adventure Today! `,
-  promoBody: `Discover the blend of traditions and modernity in Kuala Lumpur and Singapore.`,
-  callToAction: `Contact us now to secure your spot on a journey that will immerse you in the wonders of Malaysia and Singapore!.`
-};
-
-// Mauritius 
-export const mauritiusItenary = [
-  {
-    day: `Day 1`,
-    title: `Arrival in Mauritius`,
-    activities: [
-      `Welcome to the stunning island of Mauritius! Upon arrival, you'll be greeted and transferred to your exquisite accommodation. Feel the warm ocean breeze and let the relaxation begin.`
-    ]
-  },
-  {
-    day: `Day 2`,
-    title: `Beach Bliss & Water Adventures`,
-    activities: [
-      `Spend your day basking in the beauty of Mauritius' beaches. Relax under the sun, take a dip in the crystal-clear waters, or engage in exciting water sports. Your seaside retreat promises tranquility and fun.`
-    ]
-  },
-  {
-    day: `Day 3`,
-    title: `Island Exploration`,
-    activities: [
-      `Embark on an island tour to discover the enchanting landscapes of Mauritius. Visit landmarks like the Chamarel Seven Colored Earths, Grand Bassin, and the mesmerizing Chamarel Waterfall.`
-    ]
-  },
-  {
-    day: `Day 4`,
-    title: `Leisure & Culture`,
-    activities: [
-      `Indulge in a day at your leisure. Choose to unwind at the resort's spa, lounge by the pool, or explore local markets for souvenirs. In the evening, savor the island's cuisine and immerse yourself in its vibrant culture.`
-    ]
-  },
-  {
-    day: `Day 5`,
-    title: `Departure`,
-    activities: [
-      `Relish your last moments on the island. Depending on your flight time, enjoy a final stroll on the beach or take a refreshing swim. Depart Mauritius with unforgettable memories of your tropical retreat.`
-    ]
-  },
-]
-
-export const mauritiusPromoAd = {
-  promoTitle: `Book Your Mauritius Retreat Today!`,
-  promoBody: `Unveil the allure of Mauritius' beaches, landscapes, and vibrant culture.`,
-  callToAction: ` Contact us now to secure your spot on a journey that will immerse you in the blissful tranquility of Mauritius' tropical paradise!`
-};
-
-
-// China itenary
-export const chinaItenary = [
-  {
-    day: `Day 1`,
-    title: `Arrival in Guangzhou, China`,
-    activities: [
-      `Welcome to the bustling city of Guangzhou! Upon arrival, our team will ensure a smooth transfer to your comfortable accommodation. Rest up for the days of business and exploration ahead.`
-    ]
-  },
-  {
-    day: `Day 2`,
-    title: `Canton Fair Trade Exhibition Begins`,
-    activities: [
-      `Participate in the prestigious Canton Fair, where business opportunities and connections await. Engage with exhibitors, discover new products, and build valuable partnerships. Immerse yourself in the world of trade.`
-    ]
-  },
-  {
-    day: `Day 3-7`,
-    title: `Canton Fair Business Days`,
-    activities: [
-      `Dive into the business days of the Canton Fair. Network, negotiate deals, and explore the vast array of industries represented at this renowned trade exhibition.`
-    ]
-  },
-  {
-    day: `Day 8`,
-    title: `Explore Guangzhou`,
-    activities: [
-      `As the Canton Fair concludes, take the day to explore Guangzhou's cultural attractions. Visit the Chen Clan Ancestral Hall, explore the Shamian Island, and savor authentic Cantonese cuisine.`
-    ]
-  },
-  {
-    day: `Day 9`,
-    title: `Day Trip to Foshan`,
-    activities: [
-      `Embark on a day trip to Foshan, a city known for its rich history in arts and crafts. Visit the Foshan Ancestral Temple and experience the city's cultural heritage.`
-    ]
-  },
-  {
-    day: `Day 10`,
-    title: `Departure`,
-    activities: [
-      `Reflect on your Canton Fair experience and bid farewell to China. Depending on your flight time, you may have the opportunity for some last-minute exploration or shopping. Depart with new business connections and memories.`
-    ]
-  },
-]
-
-export const chinaPromoAd = {
-  promoTitle: `Book Your Canton Fair Experience Today!`,
-  promoBody: `Join the bustling world of trade at the Canton Fair and delve into the cultural richness of Guangzhou and Foshan.`,
-  callToAction: `Contact us now to secure your spot on a journey that combines business and exploration in China!`
-};
-
-// Egypt Israel
-export const egyptIsraelItenary = [
-  {
-    day: `Day 1`,
-    title: `Arrival in Cairo, Egypt`,
-    activities: [
-      `Begin your journey of faith in Cairo, Egypt's capital. Our warm welcome awaits as we transfer you to your comfortable accommodation. Take the evening to rest and reflect on the spiritual adventure ahead.`
-    ]
-  },
-  {
-    day: `Day 2`,
-    title: `Explore Ancient Egypt`,
-    activities: [
-      `Embark on a soul-enriching day of exploration. Visit the Giza Pyramids, Sphinx, and the Egyptian Museum, where artifacts from ancient civilizations await. Reflect on the connections between history and faith.`
-    ]
-  },
-  {
-    day: `Day 3`,
-    title: `Journey to Mount Sinai`,
-    activities: [
-      `Travel to the sacred Mount Sinai, where Moses received the Ten Commandments. Ascend the mountain to witness the sunrise, a moment of spiritual significance for pilgrims.`
-    ]
-  },
-  {
-    day: `Day 4`,
-    title: `Arrive in Jerusalem, Israel`,
-    activities: [
-      `Cross into Israel and arrive in Jerusalem, the heart of spirituality. Settle into your accommodation and prepare to explore the revered sites that hold deep meaning for multiple faiths.`
-    ]
-  },
-  {
-    day: `Day 5`,
-    title: `Jerusalem's Holy Sites`,
-    activities: [
-      `Embark on a profound day of visiting sacred sites. Explore the Western Wall, Church of the Holy Sepulchre, and the Dome of the Rock, reflecting on the spiritual heritage shared by many.`
-    ]
-  },
-  {
-    day: `Day 6`,
-    title: `Bethlehem and Jericho`,
-    activities: [
-      `Visit Bethlehem, the birthplace of Jesus, and explore the Church of the Nativity. Continue to Jericho, one of the world's oldest cities, and experience its historical and spiritual significance.`
-    ]
-  },
-  {
-    day: `Day 7`,
-    title: `Journey to Galilee`,
-    activities: [
-      `Travel to the serene region of Galilee, where Jesus performed many miracles. Visit sites like the Mount of Beatitudes, Tabgha, and Capernaum, connecting with the teachings of Christ.`
-    ]
-  },
-  {
-    day: `Day 8`,
-    title: `Reflection and Departure`,
-    activities: [
-      `Take a day of reflection and prayer. Depending on your flight time, you may have the opportunity to revisit a site or simply absorb the spiritual essence of your pilgrimage. Depart with a heart full of faith and memories.
-      `,
-    ]
-  },
-]
-
-export const egyptIsraelPromoAd = {
-  promoTitle: `Book Your Pilgrimage Journey Today!`,
-  promoBody: `Experience the sacredness of Egypt and Israel's sites, tracing the footsteps of prophets and connecting with centuries of faith.`,
-  callToAction: `Contact us now to secure your spot on a journey that will deepen your spiritual connection and understanding.`
-};
-
-// Paris and Barcelona
-export const parisBarcaItalyItenary = [
-  {
-    day: `Day 1`,
-    title: `Arrival in Paris, France`,
-    activities: [
-      `Begin your European adventure in the romantic city of Paris. Upon arrival, our team will warmly welcome you and transfer you to your charming accommodation. Take in the enchanting atmosphere and get ready for the days of exploration ahead.`
-    ]
-  },
-  {
-    day: `Day 2`,
-    title: `Parisian Charm`,
-    activities: [
-      `Explore the iconic landmarks of Paris. Visit the Eiffel Tower, stroll along the Seine River, and immerse yourself in the artistic wonder of the Louvre Museum. Experience the city's captivating blend of history and culture.`,
-    ]
-  },
-  {
-    day: `Day 3`,
-    title: `Travel to Barcelona, Spain`,
-    activities: [
-      `Fly to the vibrant city of Barcelona. Settle into your accommodation and spend the afternoon wandering through the charismatic streets, admiring Gaudi's architectural masterpieces, and savoring tapas.`,
-    ]
-  },
-  {
-    day: `Day 4`,
-    title: `Barcelona's Treasures`,
-    activities: [
-      `Embark on a comprehensive city tour of Barcelona. Visit the Sagrada Familia, Park Güell, and enjoy the breathtaking views from Montjuïc Hill. Feel the pulse of this creative and vibrant city.`
-    ]
-  },
-  {
-    day: `Day 5`,
-    title: `Explore Rome, Italy`,
-    activities: [
-      `Fly to the eternal city of Rome, Italy. Immerse yourself in its rich history as you visit the Colosseum, Roman Forum, and other ancient sites. Settle into your accommodation and savor authentic Italian cuisine.`
-    ]
-  },
-  {
-    day: `Day 6`,
-    title: `Vatican City & More`,
-    activities: [
-      `Experience the spiritual and artistic wonders of Vatican City. Visit St. Peter's Basilica and the Vatican Museums, home to masterpieces like Michelangelo's Sistine Chapel ceiling. Spend the afternoon exploring more of Rome's charm.`
-    ]
-  },
-  {
-    day: `Day 7`,
-    title: `Reflection and Departure`,
-    activities: [
-      `Savor your final moments in Europe, perhaps with some last-minute shopping or a leisurely stroll. Depart with cherished memories of your European journey.`,
-    ]
-  },
-]
-
-export const parisBaracItalyPromoAd = {
-  promoTitle: `Book Your European Getaway Today!`,
-  promoBody: `Unveil the beauty and diversity of Paris, Barcelona, and Rome. Limited slots are available for this exclusive offer.`,
-  callToAction: `Contact us now to secure your spot on a journey that promises to immerse you in the cultural riches of Europe!.`
-};
-
-// Bangkok
-export const bangkokItenary = [
-  {
-    day: `Day 1`,
-    title: `Arrival in Bangkok, Thailand`,
-    activities: [
-      `Begin your Thai adventure in the vibrant city of Bangkok. Upon arrival, our team will welcome you and transfer you to your comfortable accommodation. Feel the buzz of the city and prepare for an unforgettable journey.`
-    ]
-  },
-  {
-    day: `Day 2`,
-    title: `Bangkok City Exploration`,
-    activities: [
-      `Delve into Bangkok's cultural and historical treasures. Visit the Grand Palace, Wat Phra Kaew (Temple of the Emerald Buddha), and take a boat ride along the Chao Phraya River. Immerse yourself in the city's captivating blend of tradition and modernity.`,
-    ]
-  },
-  {
-    day: `Day 3`,
-    title: `Discover Bangkok's Charms`,
-    activities: [
-      `Embark on a day of exploration. Visit the bustling Damnoen Saduak Floating Market, where you can experience the vibrant local culture and sample delicious street food.`,
-    ]
-  },
-  {
-    day: `Day 4`,
-    title: `Travel to Phuket`,
-    activities: [
-      `Fly to the tropical paradise of Phuket. Settle into your beachfront accommodation and unwind under the palm trees. Enjoy the pristine beaches, turquoise waters, and laid-back island vibes.`
-    ]
-  },
-  {
-    day: `Day 5`,
-    title: `Phuket Island Adventures`,
-    activities: [
-      `Discover the natural wonders of Phuket. Explore Phi Phi Islands, snorkel through crystal-clear waters, and experience the vibrant marine life that makes these islands a paradise for underwater enthusiasts.`
-    ]
-  },
-  {
-    day: `Day 6`,
-    title: `Beach Bliss in Phuket`,
-    activities: [
-      `Spend your day soaking up the sun and enjoying the island's leisure activities. Engage in water sports, explore local markets, or simply relax on the sandy shores.`
-    ]
-  },
-  {
-    day: `Day 7`,
-    title: `Reflection and Departure`,
-    activities: [
-      `Savor your last moments in Thailand. Depending on your flight time, you may have the opportunity for a final exploration or relaxation. Depart with cherished memories of your Bangkok and Phuket escapade.`,
-    ]
-  },
-]
-export const bangkokPromoAd = {
-  promoTitle: `Unveil the excitement of Bangkok's city life and the tranquility of Phuket's beaches.`,
-  promoBody: `Unveil the beauty and diversity of Paris, Barcelona, and Rome. Limited slots are available for this exclusive offer.`,
-  callToAction: `Contact us now to secure your spot on a journey that promises a perfect balance of adventure and relaxation!`
-};
-
-// Turkey 
-export const turkeyItenary = [
-  {
-    day: `Day 1`,
-    title: `Arrival in Istanbul, Turkey`,
-    activities: [
-      `Welcome to the enchanting city of Istanbul! Upon arrival, our team will greet you and transfer you to your comfortable accommodation. Take in the blend of history and modernity that defines this vibrant city.`,
-    ]
-  },
-  {
-    day: `Day 2`,
-    title: `Explore Historical Istanbul`,
-    activities: [
-      `Embark on a captivating exploration of Istanbul's historical sites. Visit the iconic Hagia Sophia, the magnificent Blue Mosque, and the Grand Bazaar, where you can immerse yourself in the rich tapestry of Turkish culture.`,
-    ]
-  },
-  {
-    day: `Day 3`,
-    title: `Bosphorus Experience`,
-    activities: [
-      `Embark on a cruise along the Bosphorus, offering panoramic views of Istanbul's stunning skyline. Visit the opulent Dolmabahçe Palace and explore the charming neighborhoods that line the waterway.`,
-    ]
-  },
-  {
-    day: `Day 4`,
-    title: `Ottoman Heritage`,
-    activities: [
-      `Step into the opulent past of the Ottoman Empire. Visit Topkapi Palace, once the residence of Ottoman sultans, and explore the intricate details of the palace's architecture and treasures.`
-    ]
-  },
-  {
-    day: `Day 5`,
-    title: `Departure`,
-    activities: [
-      `Savor your final moments in Istanbul, perhaps with a leisurely stroll or some last-minute shopping. Depart with cherished memories of your Turkish journey and the allure of Istanbul's cultural heritage.`,
-    ]
-  },
-]
-
-export const turkeyPromoAd = {
-  promoTitle: `Unveil the excitement of Bangkok's city life and the tranquility of Phuket's beaches.`,
-  promoBody: `Embark on a journey through the heart of Turkey's cultural gem, Istanbul. Starting from just $1660 per person, this offer ensures you experience the city's rich history, vibrant markets, and iconic landmarks.`,
-  callToAction: `Contact us now to secure your spot on a journey that promises an immersive experience in the heart of Turkish culture.`
-};
